Update and rename live.py to fileless.py
