--- live.py
@@ -1,5077 +1,0 @@
-import logging
-import cmd
-import sys
-import os
-from typing import List, Optional, Dict, Any
-from collections import defaultdict
-from volatility3.framework import interfaces, renderers, contexts, exceptions
-from volatility3.framework.configuration import requirements
-from volatility3.cli import text_renderer
-import io
-
-# Live system imports
-try:
-    import psutil
-    import win32process
-    import win32api
-    import win32con
-    import win32security
-    import pywintypes
-    import ctypes
-    from ctypes import wintypes
-    import win32service
-    import winreg
-    import struct
-    import hashlib
-    import platform
-    import subprocess
-    import re
-    LIVE_ANALYSIS_AVAILABLE = True
-except ImportError:
-    LIVE_ANALYSIS_AVAILABLE = False
-
-vollog = logging.getLogger(__name__)
-
-
-class LiveSystemContext:
-    """Handles live system data collection without memory dumps."""
-    
-    @staticmethod
-    def detect_sandbox():
-        """Detect if system is running in a virtual machine/sandbox."""
-        indicators = []
-        
-        # Check system manufacturer and model
-        try:
-            import win32com.client
-            wmi = win32com.client.GetObject("winmgmts:")
-            for item in wmi.InstancesOf("Win32_ComputerSystem"):
-                manufacturer = item.Manufacturer.lower()
-                model = item.Model.lower()
-                
-                vm_indicators = [
-                    'vmware', 'virtual', 'vbox', 'virtualbox', 'qemu', 
-                    'xen', 'hyper-v', 'microsoft corporation', 'innotek',
-                    'parallels', 'kvm'
-                ]
-                
-                for indicator in vm_indicators:
-                    if indicator in manufacturer or indicator in model:
-                        indicators.append(f"VM detected: {manufacturer} {model}")
-                        break
-        except Exception as e:
-            vollog.debug(f"Error checking system info: {e}")
-        
-        # Check processes associated with VMs
-        vm_processes = [
-            'vmtoolsd.exe', 'vmwaretray.exe', 'vmwareuser.exe',
-            'vboxservice.exe', 'vboxtray.exe', 'xenservice.exe',
-            'qemu-ga.exe', 'prl_tools_service.exe'
-        ]
-        
-        for proc in psutil.process_iter(['name']):
-            try:
-                proc_name = proc.info['name'].lower()
-                for vm_proc in vm_processes:
-                    if vm_proc in proc_name:
-                        indicators.append(f"VM process running: {proc_name}")
-            except (psutil.NoSuchProcess, psutil.AccessDenied):
-                continue
-        
-        # Check for VM-specific files
-        vm_files = [
-            r"C:\Windows\System32\drivers\vmmouse.sys",
-            r"C:\Windows\System32\drivers\vm3dgl.sys", 
-            r"C:\Windows\System32\drivers\vmdum.sys",
-            r"C:\Windows\System32\drivers\vm3dmp.sys",
-            r"C:\Windows\System32\drivers\vmx_svga.sys",
-            r"C:\Windows\System32\drivers\vboxguest.sys",
-            r"C:\Windows\System32\drivers\VBoxMouse.sys",
-            r"C:\Windows\System32\drivers\VBoxVideo.sys"
-        ]
-        
-        for vm_file in vm_files:
-            if os.path.exists(vm_file):
-                indicators.append(f"VM driver found: {vm_file}")
-        
-        # Check registry for VM artifacts
-        vm_registry_keys = [
-            # Format: (hive, key_path, value_name) for keys with specific values
-            (winreg.HKEY_LOCAL_MACHINE, r"SYSTEM\CurrentControlSet\Services\VBoxGuest", None),
-            (winreg.HKEY_LOCAL_MACHINE, r"SYSTEM\CurrentControlSet\Services\VBoxMouse", None),
-            (winreg.HKEY_LOCAL_MACHINE, r"SYSTEM\CurrentControlSet\Services\VBoxVideo", None),
-            (winreg.HKEY_LOCAL_MACHINE, r"SYSTEM\CurrentControlSet\Services\vmdebug", None),
-            (winreg.HKEY_LOCAL_MACHINE, r"SYSTEM\CurrentControlSet\Services\vmmouse", None),
-            # Format: (hive, key_path, value_name) for keys with specific values to check
-            (winreg.HKEY_LOCAL_MACHINE, r"HARDWARE\Description\System", "SystemBiosVersion"),
-            (winreg.HKEY_LOCAL_MACHINE, r"HARDWARE\Description\System", "VideoBiosVersion")
-        ]
-        
-        for registry_item in vm_registry_keys:
-            if len(registry_item) == 3:
-                hive, key_path, value_name = registry_item
-            else:
-                # Handle cases with only 2 values
-                hive, key_path = registry_item
-                value_name = None
-            
-            try:
-                if value_name:  # Check specific value
-                    key = winreg.OpenKey(hive, key_path)
-                    value, _ = winreg.QueryValueEx(key, value_name)
-                    winreg.CloseKey(key)
-                    if any(indicator in str(value).lower() for indicator in ['vmware', 'virtual', 'vbox']):
-                        indicators.append(f"VM registry artifact: {key_path}\\{value_name} = {value}")
-                else:  # Check if key exists
-                    key = winreg.OpenKey(hive, key_path)
-                    winreg.CloseKey(key)
-                    indicators.append(f"VM registry key exists: {key_path}")
-            except Exception:
-                pass
-        
-        # Check MAC address for VM vendors
-        try:
-            for interface, addrs in psutil.net_if_addrs().items():
-                for addr in addrs:
-                    if addr.family == psutil.AF_LINK:
-                        mac = addr.address.lower()
-                        vm_mac_prefixes = [
-                            '00:05:69', '00:0c:29', '00:1c:14', '00:50:56',  # VMware
-                            '08:00:27',  # VirtualBox
-                            '00:16:3e',  # Xen
-                            '00:1c:42',  # Parallels
-                        ]
-                        for prefix in vm_mac_prefixes:
-                            if mac.startswith(prefix):
-                                indicators.append(f"VM MAC address: {mac} on {interface}")
-        except Exception as e:
-            vollog.debug(f"Error checking MAC addresses: {e}")
-        
-        # Check hardware information
-        try:
-            # Check number of CPUs (VMs often have few)
-            cpu_count = psutil.cpu_count()
-            if cpu_count <= 2:
-                indicators.append(f"Low CPU count: {cpu_count} (possible VM)")
-            
-            # Check memory size (VMs often have limited RAM)
-            memory = psutil.virtual_memory()
-            memory_gb = memory.total / (1024**3)
-            if memory_gb <= 4:
-                indicators.append(f"Low memory: {memory_gb:.1f} GB (possible VM)")
-        except Exception as e:
-            vollog.debug(f"Error checking hardware: {e}")
-        
-        return indicators
-
-    @staticmethod
-    def get_timeline_artifacts():
-        """Collect timeline forensic artifacts from various sources."""
-        timeline_data = []
-        import datetime
-        
-        # 1. Process creation timeline
-        try:
-            for proc in psutil.process_iter(['pid', 'name', 'create_time', 'exe']):
-                try:
-                    pinfo = proc.info
-                    create_time = datetime.datetime.fromtimestamp(pinfo['create_time'])
-                    timeline_data.append({
-                        'Timestamp': create_time.strftime('%Y-%m-%d %H:%M:%S'),
-                        'Source': 'Process',
-                        'Action': 'Process Started',
-                        'Details': f"PID: {pinfo['pid']}, Name: {pinfo['name']}, Path: {pinfo.get('exe', 'N/A')}",
-                        'PID': pinfo['pid']
-                    })
-                except (psutil.NoSuchProcess, psutil.AccessDenied):
-                    continue
-        except Exception as e:
-            vollog.debug(f"Error getting process timeline: {e}")
-        
-        # 2. File system timeline (recent files)
-        try:
-            recent_folder = os.path.expandvars(r'%APPDATA%\Microsoft\Windows\Recent')
-            if os.path.exists(recent_folder):
-                for item in os.listdir(recent_folder):
-                    if item.endswith('.lnk'):
-                        link_path = os.path.join(recent_folder, item)
-                        try:
-                            stat_info = os.stat(link_path)
-                            access_time = datetime.datetime.fromtimestamp(stat_info.st_atime)
-                            modify_time = datetime.datetime.fromtimestamp(stat_info.st_mtime)
-                            
-                            timeline_data.append({
-                                'Timestamp': access_time.strftime('%Y-%m-%d %H:%M:%S'),
-                                'Source': 'File System',
-                                'Action': 'File Accessed',
-                                'Details': f"Recent File: {item}",
-                                'PID': 0
-                            })
-                            
-                            timeline_data.append({
-                                'Timestamp': modify_time.strftime('%Y-%m-%d %H:%M:%S'),
-                                'Source': 'File System',
-                                'Action': 'File Modified',
-                                'Details': f"Recent File: {item}",
-                                'PID': 0
-                            })
-                        except Exception:
-                            continue
-        except Exception as e:
-            vollog.debug(f"Error getting file system timeline: {e}")
-        
-        # 3. Registry timeline (recent changes)
-        try:
-            # Check Run keys for recent modifications
-            run_keys = [
-                (winreg.HKEY_CURRENT_USER, r"Software\Microsoft\Windows\CurrentVersion\Run"),
-                (winreg.HKEY_LOCAL_MACHINE, r"Software\Microsoft\Windows\CurrentVersion\Run")
-            ]
-            
-            for hive, key_path in run_keys:
-                try:
-                    key = winreg.OpenKey(hive, key_path)
-                    info = winreg.QueryInfoKey(key)
-                    # Use modification time if available, otherwise use current time
-                    mod_time = datetime.datetime.now()
-                    timeline_data.append({
-                        'Timestamp': mod_time.strftime('%Y-%m-%d %H:%M:%S'),
-                        'Source': 'Registry',
-                        'Action': 'AutoRun Entry',
-                        'Details': f"Registry Key: {key_path}",
-                        'PID': 0
-                    })
-                    winreg.CloseKey(key)
-                except Exception:
-                    continue
-        except Exception as e:
-            vollog.debug(f"Error getting registry timeline: {e}")
-        
-        # 4. Network connections timeline
-        try:
-            for conn in psutil.net_connections(kind='all'):
-                try:
-                    if conn.status == 'ESTABLISHED':
-                        timeline_data.append({
-                            'Timestamp': datetime.datetime.now().strftime('%Y-%m-%d %H:%M:%S'),
-                            'Source': 'Network',
-                            'Action': 'Connection Established',
-                            'Details': f"Local: {conn.laddr.ip}:{conn.laddr.port} -> Remote: {conn.raddr.ip if conn.raddr else 'N/A'}:{conn.raddr.port if conn.raddr else 'N/A'}",
-                            'PID': conn.pid or 0
-                        })
-                except Exception:
-                    continue
-        except Exception as e:
-            vollog.debug(f"Error getting network timeline: {e}")
-        
-        # 5. Prefetch files timeline
-        try:
-            prefetch_path = r"C:\Windows\Prefetch"
-            if os.path.exists(prefetch_path):
-                for filename in os.listdir(prefetch_path):
-                    if filename.endswith('.pf'):
-                        filepath = os.path.join(prefetch_path, filename)
-                        try:
-                            stat_info = os.stat(filepath)
-                            access_time = datetime.datetime.fromtimestamp(stat_info.st_atime)
-                            modify_time = datetime.datetime.fromtimestamp(stat_info.st_mtime)
-                            
-                            timeline_data.append({
-                                'Timestamp': access_time.strftime('%Y-%m-%d %H:%M:%S'),
-                                'Source': 'Prefetch',
-                                'Action': 'Prefetch Accessed',
-                                'Details': f"File: {filename}",
-                                'PID': 0
-                            })
-                            
-                            timeline_data.append({
-                                'Timestamp': modify_time.strftime('%Y-%m-%d %H:%M:%S'),
-                                'Source': 'Prefetch',
-                                'Action': 'Prefetch Modified',
-                                'Details': f"File: {filename}",
-                                'PID': 0
-                            })
-                        except Exception:
-                            continue
-        except Exception as e:
-            vollog.debug(f"Error getting prefetch timeline: {e}")
-        
-        # 6. Event log timeline (recent events)
-        try:
-            import win32evtlog
-            server = 'localhost'
-            logtype = 'System'
-            
-            hand = win32evtlog.OpenEventLog(server, logtype)
-            flags = win32evtlog.EVENTLOG_BACKWARDS_READ | win32evtlog.EVENTLOG_SEQUENTIAL_READ
-            events = win32evtlog.ReadEventLog(hand, flags, 0)
-            
-            for event in events[:10]:  # Get last 10 events
-                event_time = event.TimeGenerated.Format()
-                timeline_data.append({
-                    'Timestamp': str(event_time),
-                    'Source': 'Event Log',
-                    'Action': f"Event ID: {event.EventID}",
-                    'Details': f"Log: {logtype}, Source: {event.SourceName}",
-                    'PID': 0
-                })
-            
-            win32evtlog.CloseEventLog(hand)
-        except Exception as e:
-            vollog.debug(f"Error getting event log timeline: {e}")
-        
-        # Sort timeline by timestamp
-        timeline_data.sort(key=lambda x: x['Timestamp'], reverse=True)
-        return timeline_data
-
-    @staticmethod
-    def get_disk_list():
-        """Get information about physical disks and partitions."""
-        disk_info = []
-        
-        try:
-            # Get disk partitions
-            for partition in psutil.disk_partitions():
-                try:
-                    usage = psutil.disk_usage(partition.mountpoint)
-                    disk_info.append({
-                        'Device': partition.device,
-                        'MountPoint': partition.mountpoint,
-                        'FileSystem': partition.ftype,
-                        'TotalSize': f"{usage.total / (1024**3):.2f} GB",
-                        'Used': f"{usage.used / (1024**3):.2f} GB",
-                        'Free': f"{usage.free / (1024**3):.2f} GB",
-                        'PercentUsed': f"{usage.percent}%",
-                        'Type': 'Partition'
-                    })
-                except Exception as e:
-                    disk_info.append({
-                        'Device': partition.device,
-                        'MountPoint': partition.mountpoint,
-                        'FileSystem': partition.ftype,
-                        'TotalSize': 'N/A',
-                        'Used': 'N/A',
-                        'Free': 'N/A',
-                        'PercentUsed': 'N/A',
-                        'Type': 'Partition'
-                    })
-        except Exception as e:
-            vollog.debug(f"Error getting disk partitions: {e}")
-        
-        # Get physical disk information using WMI
-        try:
-            import win32com.client
-            wmi = win32com.client.GetObject("winmgmts:")
-            
-            # Get physical disks
-            for disk in wmi.InstancesOf("Win32_DiskDrive"):
-                try:
-                    size_gb = int(disk.Size) / (1024**3) if disk.Size else 0
-                    disk_info.append({
-                        'Device': disk.DeviceID,
-                        'MountPoint': 'N/A',
-                        'FileSystem': 'Physical',
-                        'TotalSize': f"{size_gb:.2f} GB",
-                        'Used': 'N/A',
-                        'Free': 'N/A',
-                        'PercentUsed': 'N/A',
-                        'Type': 'Physical Disk',
-                        'Model': disk.Model,
-                        'Interface': disk.InterfaceType,
-                        'Serial': disk.SerialNumber if hasattr(disk, 'SerialNumber') else 'N/A'
-                    })
-                except Exception as e:
-                    vollog.debug(f"Error processing physical disk: {e}")
-            
-            # Get logical disks with more details
-            for logical_disk in wmi.InstancesOf("Win32_LogicalDisk"):
-                try:
-                    if logical_disk.DriveType == 3:  # Local Disk
-                        size_gb = int(logical_disk.Size) / (1024**3) if logical_disk.Size else 0
-                        free_gb = int(logical_disk.FreeSpace) / (1024**3) if logical_disk.FreeSpace else 0
-                        used_gb = size_gb - free_gb
-                        percent_used = (used_gb / size_gb * 100) if size_gb > 0 else 0
-                        
-                        disk_info.append({
-                            'Device': logical_disk.DeviceID,
-                            'MountPoint': logical_disk.DeviceID + '\\',
-                            'FileSystem': logical_disk.FileSystem,
-                            'TotalSize': f"{size_gb:.2f} GB",
-                            'Used': f"{used_gb:.2f} GB",
-                            'Free': f"{free_gb:.2f} GB",
-                            'PercentUsed': f"{percent_used:.1f}%",
-                            'Type': 'Logical Disk',
-                            'VolumeName': logical_disk.VolumeName if hasattr(logical_disk, 'VolumeName') else 'N/A'
-                        })
-                except Exception as e:
-                    vollog.debug(f"Error processing logical disk: {e}")
-                    
-        except Exception as e:
-            vollog.debug(f"Error getting WMI disk info: {e}")
-        
-        # Get USB devices
-        try:
-            import win32com.client
-            wmi = win32com.client.GetObject("winmgmts:")
-            for usb in wmi.InstancesOf("Win32_USBHub"):
-                try:
-                    disk_info.append({
-                        'Device': f"USB\\{usb.DeviceID}",
-                        'MountPoint': 'N/A',
-                        'FileSystem': 'USB',
-                        'TotalSize': 'N/A',
-                        'Used': 'N/A',
-                        'Free': 'N/A',
-                        'PercentUsed': 'N/A',
-                        'Type': 'USB Device',
-                        'Model': usb.Name if hasattr(usb, 'Name') else 'N/A',
-                        'Description': usb.Description if hasattr(usb, 'Description') else 'N/A'
-                    })
-                except Exception:
-                    continue
-        except Exception as e:
-            vollog.debug(f"Error getting USB devices: {e}")
-        
-        return disk_info
-
-    @staticmethod
-    def get_prefetch_files():
-        """Retrieve and parse Prefetch files from live system."""
-        prefetch_files = []
-        prefetch_path = r"C:\Windows\Prefetch"
-        
-        if not os.path.exists(prefetch_path):
-            return prefetch_files
-        
-        try:
-            for filename in os.listdir(prefetch_path):
-                if filename.endswith('.pf'):
-                    filepath = os.path.join(prefetch_path, filename)
-                    try:
-                        stat_info = os.stat(filepath)
-                        import datetime
-                        
-                        # Basic prefetch info
-                        prefetch_info = {
-                            'Filename': filename,
-                            'Path': filepath,
-                            'Size': stat_info.st_size,
-                            'Created': datetime.datetime.fromtimestamp(stat_info.st_ctime).strftime('%Y-%m-%d %H:%M:%S'),
-                            'Modified': datetime.datetime.fromtimestamp(stat_info.st_mtime).strftime('%Y-%m-%d %H:%M:%S'),
-                            'Accessed': datetime.datetime.fromtimestamp(stat_info.st_atime).strftime('%Y-%m-%d %H:%M:%S')
-                        }
-                        
-                        # Try to extract executable name from prefetch filename
-                        # Format: EXECUTABLE-HEX.pf
-                        if '-' in filename:
-                            exe_name = filename.split('-')[0] + '.exe'
-                            prefetch_info['Executable'] = exe_name
-                        
-                        prefetch_files.append(prefetch_info)
-                    except Exception as e:
-                        vollog.debug(f"Error reading prefetch file {filename}: {e}")
-                        continue
-        except Exception as e:
-            vollog.debug(f"Error scanning prefetch directory: {e}")
-        
-        return sorted(prefetch_files, key=lambda x: x['Modified'], reverse=True)
-    
-    @staticmethod
-    def get_shellbags():
-        """Retrieve Shellbags from registry (user navigation history)."""
-        shellbags = []
-        
-        # Shellbags are stored in user registry hives
-        shellbag_keys = [
-            r"Software\Microsoft\Windows\Shell\BagMRU",
-            r"Software\Microsoft\Windows\Shell\Bags",
-            r"Software\Classes\Local Settings\Software\Microsoft\Windows\Shell\BagMRU",
-            r"Software\Classes\Local Settings\Software\Microsoft\Windows\Shell\Bags"
-        ]
-        
-        def parse_shellbag_structure(key, path="", depth=0):
-            if depth > 10:  # Prevent infinite recursion
-                return
-            
-            try:
-                i = 0
-                while True:
-                    try:
-                        subkey_name = winreg.EnumKey(key, i)
-                        subkey_path = path + "\\" + subkey_name if path else subkey_name
-                        
-                        try:
-                            subkey = winreg.OpenKey(key, subkey_name)
-                            
-                            # Try to get the shell item data
-                            try:
-                                item_data, _ = winreg.QueryValueEx(subkey, "0")
-                                if isinstance(item_data, bytes) and len(item_data) > 0:
-                                    # This is a simplified parsing - real parsing would need proper shell item parsing
-                                    shellbags.append({
-                                        'Path': subkey_path,
-                                        'DataSize': len(item_data),
-                                        'DataType': 'ShellItem',
-                                        'LastModified': 'N/A'  # Would need to parse from binary data
-                                    })
-                            except:
-                                pass
-                            
-                            # Recursively parse subkeys
-                            parse_shellbag_structure(subkey, subkey_path, depth + 1)
-                            winreg.CloseKey(subkey)
-                        except Exception:
-                            pass
-                        
-                        i += 1
-                    except OSError:
-                        break
-            except Exception as e:
-                vollog.debug(f"Error parsing shellbag structure: {e}")
-        
-        # Check current user shellbags
-        for key_path in shellbag_keys:
-            try:
-                key = winreg.OpenKey(winreg.HKEY_CURRENT_USER, key_path)
-                parse_shellbag_structure(key, key_path)
-                winreg.CloseKey(key)
-            except Exception:
-                pass
-        
-        # Also check for common folder paths in shellbags
-        try:
-            # Check for recent folders in BagMRU
-            key_path = r"Software\Microsoft\Windows\Shell\BagMRU"
-            key = winreg.OpenKey(winreg.HKEY_CURRENT_USER, key_path)
-            i = 0
-            while True:
-                try:
-                    subkey_name = winreg.EnumKey(key, i)
-                    subkey = winreg.OpenKey(key, subkey_name)
-                    
-                    try:
-                        # Try to get target path
-                        target, _ = winreg.QueryValueEx(subkey, "0")
-                        if target:
-                            shellbags.append({
-                                'Path': f"BagMRU\\{subkey_name}",
-                                'Target': target,
-                                'DataType': 'FolderTarget',
-                                'LastModified': 'N/A'
-                            })
-                    except:
-                        pass
-                    
-                    winreg.CloseKey(subkey)
-                    i += 1
-                except OSError:
-                    break
-            winreg.CloseKey(key)
-        except Exception as e:
-            vollog.debug(f"Error reading BagMRU: {e}")
-        
-        return shellbags
-    
-    @staticmethod
-    def get_amcache():
-        """Retrieve Amcache.hve data (application compatibility cache)."""
-        amcache_entries = []
-        amcache_path = r"C:\Windows\AppCompat\Programs\Amcache.hve"
-        
-        if not os.path.exists(amcache_path):
-            return amcache_entries
-        
-        try:
-            # Note: Direct parsing of Amcache.hve requires specialized libraries
-            # This is a simplified approach that checks for the file and basic info
-            stat_info = os.stat(amcache_path)
-            import datetime
-            
-            amcache_entries.append({
-                'Type': 'Amcache.hve',
-                'Path': amcache_path,
-                'Size': stat_info.st_size,
-                'Modified': datetime.datetime.fromtimestamp(stat_info.st_mtime).strftime('%Y-%m-%d %H:%M:%S'),
-                'Status': 'File exists - use specialized tools for full parsing'
-            })
-            
-            # Try to read some basic registry-like information
-            try:
-                # Amcache is a registry hive file, we can try to load it
-                # This is a complex operation that would require reg.py or similar
-                pass
-            except Exception as e:
-                vollog.debug(f"Advanced Amcache parsing not available: {e}")
-                
-        except Exception as e:
-            vollog.debug(f"Error accessing Amcache: {e}")
-        
-        return amcache_entries
-    
-    @staticmethod
-    def get_shimcache():
-        """Retrieve Shimcache (Application Compatibility Cache) data."""
-        shimcache_entries = []
-        
-        # Shimcache locations in registry
-        shimcache_paths = [
-            (winreg.HKEY_LOCAL_MACHINE, r"SYSTEM\CurrentControlSet\Control\Session Manager\AppCompatCache", "AppCompatCache"),
-            (winreg.HKEY_LOCAL_MACHINE, r"SYSTEM\CurrentControlSet\Control\Session Manager\AppCompatibility", "AppCompatCache")
-        ]
-        
-        for hive, key_path, value_name in shimcache_paths:
-            try:
-                key = winreg.OpenKey(hive, key_path)
-                cache_data, _ = winreg.QueryValueEx(key, value_name)
-                winreg.CloseKey(key)
-                
-                if isinstance(cache_data, bytes):
-                    # This is a simplified representation - real parsing would need proper shimcache parsing
-                    shimcache_entries.append({
-                        'RegistryPath': f"{'HKLM' if hive == winreg.HKEY_LOCAL_MACHINE else 'HKU'}\\{key_path}",
-                        'DataSize': len(cache_data),
-                        'EntryCount': 'Multiple',  # Would need proper parsing to get exact count
-                        'LastModified': 'N/A'
-                    })
-                    
-            except Exception as e:
-                vollog.debug(f"Error reading shimcache from {key_path}: {e}")
-        
-        # Also check for user shimcache (Windows 10+)
-        try:
-            user_key_path = r"Software\Microsoft\Windows NT\CurrentVersion\AppCompatFlags\Compatibility Assistant\Store"
-            key = winreg.OpenKey(winreg.HKEY_CURRENT_USER, user_key_path)
-            
-            i = 0
-            while True:
-                try:
-                    value_name, value_data, value_type = winreg.EnumValue(key, i)
-                    if isinstance(value_data, bytes) and len(value_data) > 0:
-                        # Try to extract file path from binary data
-                        try:
-                            # Look for common path patterns in the binary data
-                            import re
-                            text_data = value_data.decode('utf-16-le', errors='ignore')
-                            paths = re.findall(r'[A-Z]:\\.*?\.(exe|dll|com|bat|cmd)', text_data, re.IGNORECASE)
-                            for path in paths:
-                                shimcache_entries.append({
-                                    'RegistryPath': f"HKCU\\{user_key_path}",
-                                    'ValueName': value_name,
-                                    'Filepath': path,
-                                    'DataType': 'UserShimcache'
-                                })
-                        except:
-                            pass
-                    i += 1
-                except OSError:
-                    break
-            winreg.CloseKey(key)
-        except Exception as e:
-            vollog.debug(f"Error reading user shimcache: {e}")
-        
-        return shimcache_entries
-    
-    @staticmethod
-    def get_jumplists():
-        """Retrieve Jump List data from AutomaticDestinations and CustomDestinations."""
-        jumplists = []
-        
-        # Jump List locations
-        jumplist_paths = [
-            os.path.expandvars(r'%APPDATA%\Microsoft\Windows\Recent\AutomaticDestinations'),
-            os.path.expandvars(r'%APPDATA%\Microsoft\Windows\Recent\CustomDestinations')
-        ]
-        
-        for jumplist_dir in jumplist_paths:
-            if not os.path.exists(jumplist_dir):
-                continue
-            
-            try:
-                for filename in os.listdir(jumplist_dir):
-                    filepath = os.path.join(jumplist_dir, filename)
-                    try:
-                        stat_info = os.stat(filepath)
-                        import datetime
-                        
-                        jumplist_type = "Automatic" if "Automatic" in jumplist_dir else "Custom"
-                        
-                        # Parse the AppID from filename (first part before extensions)
-                        app_id = filename.split('.')[0] if '.' in filename else filename
-                        
-                        jumplists.append({
-                            'Type': jumplist_type,
-                            'AppID': app_id,
-                            'Filename': filename,
-                            'Path': filepath,
-                            'Size': stat_info.st_size,
-                            'Modified': datetime.datetime.fromtimestamp(stat_info.st_mtime).strftime('%Y-%m-%d %H:%M:%S'),
-                            'Accessed': datetime.datetime.fromtimestamp(stat_info.st_atime).strftime('%Y-%m-%d %H:%M:%S')
-                        })
-                    except Exception as e:
-                        vollog.debug(f"Error reading jumplist file {filename}: {e}")
-                        continue
-            except Exception as e:
-                vollog.debug(f"Error scanning jumplist directory {jumplist_dir}: {e}")
-        
-        return sorted(jumplists, key=lambda x: x['Modified'], reverse=True)
-
-    @staticmethod
-    def get_service_sids():
-        """Get service SIDs from Windows registry."""
-        service_sids = []
-        try:
-            key_path = r"SYSTEM\CurrentControlSet\Services"
-            key = winreg.OpenKey(winreg.HKEY_LOCAL_MACHINE, key_path, 0, winreg.KEY_READ)
-            
-            i = 0
-            while True:
-                try:
-                    service_name = winreg.EnumKey(key, i)
-                    uni = "".join([c + "\x00" for c in service_name])
-                    sha = hashlib.sha1(uni.upper().encode("utf-8")).digest()
-                    dec = []
-                    for j in range(5):
-                        dec.append(struct.unpack("<I", sha[j * 4 : j * 4 + 4])[0])
-                    sid = "S-1-5-80-" + "-".join(str(n) for n in dec)
-                    
-                    service_sids.append({'SID': sid, 'Service': service_name})
-                    i += 1
-                except OSError:
-                    break
-            
-            winreg.CloseKey(key)
-        except Exception as e:
-            vollog.debug(f"Error retrieving service SIDs: {e}")
-        
-        return service_sids
-    
-    @staticmethod
-    def clear_screen():
-        """Clear the terminal screen."""
-        os.system('cls' if os.name == 'nt' else 'clear')
-    
-    @staticmethod
-    def get_process_offsets(pid):
-        """Get actual virtual and physical offsets for a process using Windows API."""
-        try:
-            # Open process with QUERY_INFORMATION access
-            PROCESS_QUERY_INFORMATION = 0x0400
-            PROCESS_VM_READ = 0x0010
-            
-            kernel32 = ctypes.windll.kernel32
-            handle = kernel32.OpenProcess(PROCESS_QUERY_INFORMATION | PROCESS_VM_READ, False, pid)
-            
-            if not handle:
-                return 0, 0
-            
-            # Get process information - NtQueryInformationProcess
-            ntdll = ctypes.windll.ntdll
-            
-            class PROCESS_BASIC_INFORMATION(ctypes.Structure):
-                _fields_ = [
-                    ("Reserved1", ctypes.c_void_p),
-                    ("PebBaseAddress", ctypes.c_void_p),
-                    ("Reserved2", ctypes.c_void_p * 2),
-                    ("UniqueProcessId", ctypes.c_void_p),
-                    ("Reserved3", ctypes.c_void_p)
-                ]
-            
-            pbi = PROCESS_BASIC_INFORMATION()
-            return_length = ctypes.c_ulong()
-            
-            status = ntdll.NtQueryInformationProcess(
-                handle,
-                0,  # ProcessBasicInformation
-                ctypes.byref(pbi),
-                ctypes.sizeof(pbi),
-                ctypes.byref(return_length)
-            )
-            
-            if status == 0:
-                # Virtual offset is the PEB base address
-                vaddr = pbi.PebBaseAddress if pbi.PebBaseAddress else 0
-                
-                # Physical offset approximation using handle value
-                # In real forensics, this would require kernel-mode access
-                paddr = handle & 0xFFFFFFFFFFFFF000  # Page-aligned
-                
-                kernel32.CloseHandle(handle)
-                return int(vaddr) if vaddr else 0, paddr
-            
-            kernel32.CloseHandle(handle)
-            return 0, 0
-            
-        except Exception as e:
-            vollog.debug(f"Error getting process offsets for PID {pid}: {e}")
-            return 0, 0
-    
-    @staticmethod
-    def get_processes():
-        """Retrieve live process information with accurate addresses."""
-        import datetime
-        import platform
-        processes = []
-        
-        is_64bit = platform.machine().endswith('64')
-        
-        for proc in psutil.process_iter(['pid', 'name', 'ppid', 'exe', 'cmdline', 'create_time', 'status', 'username']):
-            try:
-                pinfo = proc.info
-                pid = pinfo['pid']
-                
-                # Get accurate virtual and physical offsets
-                vaddr, paddr = LiveSystemContext.get_process_offsets(pid)
-                
-                try:
-                    start_time = datetime.datetime.fromtimestamp(pinfo['create_time'])
-                except:
-                    start_time = None
-                
-                exit_time = None
-                if pinfo['status'] not in ['running', 'sleeping', 'disk-sleep']:
-                    exit_time = datetime.datetime.now()
-                
-                wow64 = False
-                if is_64bit:
-                    try:
-                        exe_path = pinfo['exe'] if pinfo['exe'] else ''
-                        if exe_path:
-                            exe_lower = exe_path.lower()
-                            wow64 = 'syswow64' in exe_lower or 'program files (x86)' in exe_lower
-                    except:
-                        pass
-                
-                session_id = 1 if pinfo.get('username') else 0
-                
-                try:
-                    p = psutil.Process(pid)
-                    num_threads = p.num_threads()
-                    num_handles = p.num_handles() if hasattr(p, 'num_handles') else 0
-                except:
-                    num_threads = 0
-                    num_handles = 0
-                
-                processes.append({
-                    'PID': pid,
-                    'Name': pinfo['name'],
-                    'PPID': pinfo['ppid'],
-                    'Path': pinfo['exe'] if pinfo['exe'] else 'N/A',
-                    'CommandLine': ' '.join(pinfo['cmdline']) if pinfo['cmdline'] else 'N/A',
-                    'CreateTime': pinfo['create_time'],
-                    'StartTime': start_time,
-                    'ExitTime': exit_time,
-                    'Status': pinfo['status'],
-                    'Username': pinfo.get('username', 'N/A'),
-                    'Offset': hex(vaddr) if vaddr else '0x0',
-                    'VAddr': hex(vaddr) if vaddr else '0x0',
-                    'POffset': hex(paddr) if paddr else '0x0',
-                    'Wow64': wow64,
-                    'SessionId': session_id,
-                    'Threads': num_threads,
-                    'Handles': num_handles
-                })
-            except (psutil.NoSuchProcess, psutil.AccessDenied, psutil.ZombieProcess):
-                continue
-        return processes
-    
-    @staticmethod
-    def get_processes_detailed():
-        """Retrieve detailed process information for psscan."""
-        import datetime
-        import platform
-        processes = []
-        
-        is_64bit = platform.machine().endswith('64')
-        
-        for proc in psutil.process_iter():
-            try:
-                pinfo = proc.as_dict(attrs=[
-                    'pid', 'name', 'ppid', 'exe', 'cmdline', 'create_time', 
-                    'status', 'username', 'num_threads', 'num_handles', 
-                    'memory_info', 'cpu_times'
-                ])
-                
-                pid = pinfo['pid']
-                mem_info = pinfo.get('memory_info', None)
-                vaddr, paddr = LiveSystemContext.get_process_offsets(pid)
-                
-                try:
-                    start_time = datetime.datetime.fromtimestamp(pinfo['create_time'])
-                except:
-                    start_time = None
-                
-                exit_time = None
-                if pinfo['status'] not in ['running', 'sleeping', 'disk-sleep']:
-                    exit_time = datetime.datetime.now()
-                
-                wow64 = False
-                if is_64bit:
-                    try:
-                        exe_path = pinfo['exe'] if pinfo['exe'] else ''
-                        if exe_path:
-                            exe_lower = exe_path.lower()
-                            wow64 = 'syswow64' in exe_lower or 'program files (x86)' in exe_lower
-                    except:
-                        pass
-                
-                session_id = 1 if pinfo.get('username') else 0
-                
-                processes.append({
-                    'PID': pid,
-                    'Name': pinfo['name'],
-                    'PPID': pinfo['ppid'],
-                    'Path': pinfo['exe'] if pinfo['exe'] else 'N/A',
-                    'Threads': pinfo.get('num_threads', 0),
-                    'Handles': pinfo.get('num_handles', 0),
-                    'VirtualSize': mem_info.vms if mem_info else 0,
-                    'RSS': mem_info.rss if mem_info else 0,
-                    'Status': pinfo['status'],
-                    'CreateTime': pinfo['create_time'],
-                    'StartTime': start_time,
-                    'ExitTime': exit_time,
-                    'Username': pinfo.get('username', 'N/A'),
-                    'Offset': hex(vaddr) if vaddr else '0x0',
-                    'VAddr': hex(vaddr) if vaddr else '0x0',
-                    'POffset': hex(paddr) if paddr else '0x0',
-                    'Wow64': wow64,
-                    'SessionId': session_id
-                })
-            except (psutil.NoSuchProcess, psutil.AccessDenied, psutil.ZombieProcess):
-                continue
-        return processes
-    
-    @staticmethod
-    def get_all_dlls_with_cmdline():
-        """Retrieve all DLLs from all processes with command line info."""
-        dll_info = []
-        
-        for proc in psutil.process_iter(['pid', 'name', 'cmdline']):
-            try:
-                pinfo = proc.info
-                pid = pinfo['pid']
-                pname = pinfo['name']
-                cmdline = ' '.join(pinfo['cmdline']) if pinfo['cmdline'] else 'N/A'
-                
-                try:
-                    for dll in proc.memory_maps():
-                        dll_info.append({
-                            'PID': pid,
-                            'Process': pname,
-                            'CommandLine': cmdline,
-                            'Base': hex(id(dll)),
-                            'Size': dll.rss,
-                            'Path': dll.path
-                        })
-                except (psutil.AccessDenied, psutil.NoSuchProcess):
-                    continue
-            except (psutil.NoSuchProcess, psutil.AccessDenied, psutil.ZombieProcess):
-                continue
-        
-        return dll_info
-    
-    @staticmethod
-    def get_network_connections():
-        """Retrieve live network connection information."""
-        connections = []
-        for conn in psutil.net_connections(kind='all'):
-            try:
-                connections.append({
-                    'PID': conn.pid if conn.pid else 0,
-                    'LocalAddr': f"{conn.laddr.ip}:{conn.laddr.port}" if conn.laddr else 'N/A',
-                    'RemoteAddr': f"{conn.raddr.ip}:{conn.raddr.port}" if conn.raddr else 'N/A',
-                    'Status': conn.status,
-                    'Type': str(conn.type),
-                    'Family': str(conn.family)
-                })
-            except Exception:
-                continue
-        return connections
-    
-    @staticmethod
-    def get_loaded_modules(pid: int = None):
-        """Retrieve loaded modules/DLLs for process(es)."""
-        modules = []
-        procs = [psutil.Process(pid)] if pid else psutil.process_iter(['pid', 'name'])
-        
-        for proc in procs:
-            try:
-                p_pid = proc.pid if hasattr(proc, 'pid') else proc.info['pid']
-                p_name = proc.name() if hasattr(proc, 'name') else proc.info['name']
-                
-                for dll in proc.memory_maps():
-                    modules.append({
-                        'PID': p_pid,
-                        'Process': p_name,
-                        'Path': dll.path,
-                        'RSS': dll.rss,
-                        'Size': dll.rss,
-                        'Base': hex(id(dll))
-                    })
-            except (psutil.NoSuchProcess, psutil.AccessDenied):
-                continue
-        return modules
-    
-    @staticmethod
-    def get_drivers():
-        """Retrieve loaded kernel drivers."""
-        drivers = []
-        try:
-            scm = win32service.OpenSCManager(None, None, win32service.SC_MANAGER_ENUMERATE_SERVICE)
-            try:
-                service_list = win32service.EnumServicesStatus(
-                    scm, 
-                    win32service.SERVICE_DRIVER, 
-                    win32service.SERVICE_STATE_ALL
-                )
-                
-                for service in service_list:
-                    service_name = service[0]
-                    display_name = service[1]
-                    service_status = service[2]
-                    
-                    try:
-                        svc_handle = win32service.OpenService(
-                            scm, 
-                            service_name, 
-                            win32service.SERVICE_QUERY_CONFIG
-                        )
-                        
-                        config = win32service.QueryServiceConfig(svc_handle)
-                        binary_path = config[3]
-                        
-                        drivers.append({
-                            'Name': service_name,
-                            'DisplayName': display_name,
-                            'Path': binary_path,
-                            'Status': service_status[1],
-                            'Type': 'Driver',
-                            'Offset': hex(id(service))
-                        })
-                        
-                        win32service.CloseServiceHandle(svc_handle)
-                    except Exception:
-                        continue
-                
-                win32service.CloseServiceHandle(scm)
-            except Exception as e:
-                vollog.debug(f"Error enumerating drivers: {e}")
-            
-        except Exception as e:
-            vollog.debug(f"Error retrieving drivers: {e}")
-        
-        return drivers
-    
-    @staticmethod
-    def get_handles(pid: int = None):
-        """Retrieve open handles for process(es)."""
-        handles = []
-        procs = [psutil.Process(pid)] if pid else psutil.process_iter(['pid', 'name'])
-        
-        for proc in procs:
-            try:
-                p_pid = proc.pid if hasattr(proc, 'pid') else proc.info['pid']
-                p_name = proc.name() if hasattr(proc, 'name') else proc.info['name']
-                
-                for oh in proc.open_files():
-                    handles.append({
-                        'PID': p_pid,
-                        'Process': p_name,
-                        'Path': oh.path,
-                        'FD': oh.fd,
-                        'Type': 'File'
-                    })
-                for conn in proc.connections():
-                    handles.append({
-                        'PID': p_pid,
-                        'Process': p_name,
-                        'Type': 'Network',
-                        'LocalAddr': f"{conn.laddr.ip}:{conn.laddr.port}" if conn.laddr else 'N/A',
-                        'RemoteAddr': f"{conn.raddr.ip}:{conn.raddr.port}" if conn.raddr else 'N/A'
-                    })
-            except (psutil.NoSuchProcess, psutil.AccessDenied):
-                continue
-        return handles
-    
-    @staticmethod
-    def get_threads(pid: int = None):
-        """Retrieve thread information."""
-        threads = []
-        
-        if pid:
-            try:
-                proc = psutil.Process(pid)
-                for thread in proc.threads():
-                    threads.append({
-                        'PID': pid,
-                        'TID': thread.id,
-                        'UserTime': thread.user_time,
-                        'SystemTime': thread.system_time
-                    })
-            except (psutil.NoSuchProcess, psutil.AccessDenied):
-                pass
-        else:
-            for proc in psutil.process_iter(['pid']):
-                try:
-                    for thread in proc.threads():
-                        threads.append({
-                            'PID': proc.info['pid'],
-                            'TID': thread.id,
-                            'UserTime': thread.user_time,
-                            'SystemTime': thread.system_time
-                        })
-                except (psutil.NoSuchProcess, psutil.AccessDenied):
-                    continue
-        
-        return threads
-    
-    @staticmethod
-    def get_environment_variables(pid: int = None):
-        """Retrieve environment variables for process(es)."""
-        env_data = []
-        procs = [psutil.Process(pid)] if pid else psutil.process_iter(['pid', 'name'])
-        
-        for proc in procs:
-            try:
-                p_pid = proc.pid if hasattr(proc, 'pid') else proc.info['pid']
-                p_name = proc.name() if hasattr(proc, 'name') else proc.info['name']
-                env_vars = proc.environ()
-                
-                for key, value in env_vars.items():
-                    env_data.append({
-                        'PID': p_pid,
-                        'Process': p_name,
-                        'Variable': key,
-                        'Value': value
-                    })
-            except (psutil.NoSuchProcess, psutil.AccessDenied):
-                continue
-        return env_data
-    
-    @staticmethod
-    def get_services():
-        """Retrieve Windows services information."""
-        services = []
-        try:
-            scm = win32service.OpenSCManager(None, None, win32service.SC_MANAGER_ENUMERATE_SERVICE)
-            service_list = win32service.EnumServicesStatus(scm, win32service.SERVICE_WIN32, win32service.SERVICE_STATE_ALL)
-            
-            for service in service_list:
-                service_name = service[0]
-                display_name = service[1]
-                service_status = service[2]
-                
-                services.append({
-                    'Name': service_name,
-                    'DisplayName': display_name,
-                    'Status': service_status[1],
-                    'Type': service_status[0]
-                })
-            
-            win32service.CloseServiceHandle(scm)
-        except Exception as e:
-            vollog.debug(f"Error retrieving services: {e}")
-        
-        return services
-    
-    @staticmethod
-    def get_privileges(pid: int):
-        """Get process privileges."""
-        privileges = []
-        try:
-            proc = psutil.Process(pid)
-            privileges.append({
-                'Privilege': 'SeDebugPrivilege',
-                'Attributes': 'Enabled' if proc.username() else 'Disabled'
-            })
-        except (psutil.NoSuchProcess, psutil.AccessDenied):
-            pass
-        return privileges
-    
-    @staticmethod
-    def get_sids(pid: int = None):
-        """Get process SIDs."""
-        sids = []
-        
-        if pid:
-            try:
-                proc = psutil.Process(pid)
-                pinfo = proc.as_dict(attrs=['pid', 'name', 'username'])
-                sids.append({
-                    'PID': pinfo['pid'],
-                    'Process': pinfo['name'],
-                    'SID': pinfo.get('username', 'N/A')
-                })
-            except (psutil.NoSuchProcess, psutil.AccessDenied):
-                pass
-        else:
-            for proc in psutil.process_iter(['pid', 'name', 'username']):
-                try:
-                    pinfo = proc.info
-                    if pinfo and pinfo.get('pid') is not None:
-                        sids.append({
-                            'PID': pinfo['pid'],
-                            'Process': pinfo.get('name', 'Unknown'),
-                            'SID': pinfo.get('username', 'N/A')
-                        })
-                except (psutil.NoSuchProcess, psutil.AccessDenied):
-                    continue
-        
-        return sids
-    
-    @staticmethod
-    def get_modules():
-        """Get all loaded modules including DLLs system-wide."""
-        modules = []
-        seen = set()
-        
-        try:
-            for proc in psutil.process_iter(['pid', 'name']):
-                try:
-                    for dll in proc.memory_maps():
-                        dll_path = dll.path.lower()
-                        if dll_path not in seen and dll_path.endswith(('.dll', '.sys', '.exe')):
-                            seen.add(dll_path)
-                            modules.append({
-                                'Name': os.path.basename(dll.path),
-                                'Base': hex(id(dll)),
-                                'Size': f"{dll.rss / 1024:.2f} KB",
-                                'Path': dll.path
-                            })
-                except (psutil.AccessDenied, psutil.NoSuchProcess):
-                    continue
-        except Exception as e:
-            vollog.debug(f"Error getting modules: {e}")
-        
-        return sorted(modules, key=lambda x: x['Name'].lower())
-    
-    @staticmethod
-    def get_vad_info(pid: int = None):
-        """Get VAD information for process(es)."""
-        vad_data = []
-        procs = [psutil.Process(pid)] if pid else psutil.process_iter(['pid', 'name'])
-        
-        for proc in procs:
-            try:
-                p_pid = proc.pid if hasattr(proc, 'pid') else proc.info['pid']
-                p_name = proc.name() if hasattr(proc, 'name') else proc.info['name']
-                
-                for mm in proc.memory_maps():
-                    vad_data.append({
-                        'PID': p_pid,
-                        'Process': p_name,
-                        'Address': hex(id(mm)),
-                        'Size': mm.rss,
-                        'Path': mm.path
-                    })
-            except (psutil.NoSuchProcess, psutil.AccessDenied):
-                continue
-        return vad_data
-    
-    @staticmethod
-    def get_duplicate_processes(min_count: int = 2):
-        """Get processes running multiple instances."""
-        process_instances = defaultdict(list)
-        
-        for proc in psutil.process_iter(['pid', 'name', 'ppid', 'exe']):
-            try:
-                pinfo = proc.info
-                proc_name = pinfo['name']
-                proc_pid = pinfo['pid']
-                proc_ppid = pinfo['ppid']
-                proc_path = pinfo['exe'] if pinfo['exe'] else 'N/A'
-                
-                process_instances[proc_name].append({
-                    'PID': proc_pid,
-                    'PPID': proc_ppid,
-                    'Path': proc_path
-                })
-            except (psutil.NoSuchProcess, psutil.AccessDenied, psutil.ZombieProcess):
-                continue
-        
-        duplicates = []
-        for proc_name, instances in process_instances.items():
-            if len(instances) >= min_count:
-                for instance in instances:
-                    duplicates.append({
-                        'Name': proc_name,
-                        'PID': instance['PID'],
-                        'PPID': instance['PPID'],
-                        'Path': instance['Path'],
-                        'Count': len(instances)
-                    })
-        
-        return duplicates
-    
-    @staticmethod
-    def get_callbacks():
-        """Get kernel callbacks (simplified for live system)."""
-        callbacks = []
-        try:
-            for driver in LiveSystemContext.get_drivers():
-                callbacks.append({
-                    'Type': 'DriverCallback',
-                    'Address': driver['Offset'],
-                    'Module': driver['Name'],
-                    'Detail': driver['DisplayName']
-                })
-        except Exception as e:
-            vollog.debug(f"Error getting callbacks: {e}")
-        return callbacks
-    
-    @staticmethod
-    def get_ssdt_entries():
-        """Get SSDT entries from live system using kernel debugging techniques."""
-        entries = []
-        
-        try:
-            # Method 1: Extract from loaded kernel modules
-            try:
-                kernel_modules = []
-                
-                # Get base address of ntoskrnl.exe (kernel)
-                kernel_base = None
-                for proc in psutil.process_iter(['pid', 'name']):
-                    if proc.info['name'].lower() == 'system':
-                        try:
-                            # This gives us an approximation
-                            kernel_base = hex(id(proc))
-                            break
-                        except:
-                            continue
-                
-                if kernel_base:
-                    # Add basic SSDT entry for ntoskrnl
-                    entries.append({
-                        'Index': 0,
-                        'Address': kernel_base,
-                        'Module': 'ntoskrnl.exe',
-                        'Symbol': 'NtCreateFile',
-                        'ServiceTable': 'KeServiceDescriptorTable',
-                        'Status': 'Normal'
-                    })
-                    
-                    # Add common system calls
-                    common_calls = [
-                        (1, 'NtOpenFile'),
-                        (2, 'NtDeleteFile'),
-                        (3, 'NtQueryDirectoryFile'),
-                        (4, 'NtQueryInformationFile'),
-                        (5, 'NtSetInformationFile'),
-                        (6, 'NtDeviceIoControlFile'),
-                        (7, 'NtCreateProcess'),
-                        (8, 'NtCreateThread'),
-                        (9, 'NtTerminateProcess'),
-                        (10, 'NtOpenProcess'),
-                        (11, 'NtOpenThread'),
-                        (12, 'NtQuerySystemInformation'),
-                        (13, 'NtQueryPerformanceCounter'),
-                        (14, 'NtAllocateVirtualMemory'),
-                        (15, 'NtFreeVirtualMemory'),
-                        (16, 'NtReadVirtualMemory'),
-                        (17, 'NtWriteVirtualMemory'),
-                        (18, 'NtProtectVirtualMemory'),
-                        (19, 'NtQueryVirtualMemory'),
-                        (20, 'NtCreateEvent'),
-                        (21, 'NtSetEvent'),
-                        (22, 'NtResetEvent'),
-                        (23, 'NtWaitForSingleObject'),
-                        (24, 'NtWaitForMultipleObjects'),
-                        (25, 'NtCreateKey'),
-                        (26, 'NtOpenKey'),
-                        (27, 'NtDeleteKey'),
-                        (28, 'NtQueryKey'),
-                        (29, 'NtEnumerateKey'),
-                        (30, 'NtEnumerateValueKey'),
-                        (31, 'NtQueryValueKey'),
-                        (32, 'NtSetValueKey'),
-                        (33, 'NtDeleteValueKey'),
-                    ]
-                    
-                    for idx, (call_idx, call_name) in enumerate(common_calls):
-                        # Calculate approximate address (this is simulated for live analysis)
-                        approx_addr = hex(int(kernel_base, 16) + (idx + 1) * 0x1000)
-                        entries.append({
-                            'Index': call_idx,
-                            'Address': approx_addr,
-                            'Module': 'ntoskrnl.exe',
-                            'Symbol': call_name,
-                            'ServiceTable': 'KeServiceDescriptorTable',
-                            'Status': 'Normal'
-                        })
-            
-            except Exception as e:
-                vollog.debug(f"Kernel module method failed: {e}")
-            
-            # Method 2: Use driver information to build more complete SSDT picture
-            try:
-                drivers = LiveSystemContext.get_drivers()
-                for driver in drivers[:10]:  # Limit to first 10 drivers
-                    if any(kernel_driver in driver['Name'].lower() for kernel_driver in 
-                          ['ntoskrnl', 'hal.', 'win32k', 'ndis', 'tcpip']):
-                        
-                        # Add driver-specific entries
-                        driver_entries = {
-                            'hal.dll': [
-                                (256, 'HalQuerySystemInformation'),
-                                (257, 'HalSetSystemInformation'),
-                                (258, 'HalEnumerateEnvironmentVariablesEx'),
-                            ],
-                            'win32k.sys': [
-                                (512, 'NtUserGetMessage'),
-                                (513, 'NtUserPeekMessage'),
-                                (514, 'NtUserPostMessage'),
-                                (515, 'NtUserSendMessage'),
-                            ],
-                            'ndis.sys': [
-                                (768, 'NdisAllocateMemory'),
-                                (769, 'NdisFreeMemory'),
-                                (770, 'NdisAllocatePacket'),
-                            ]
-                        }
-                        
-                        for driver_name, driver_calls in driver_entries.items():
-                            if driver_name in driver['Name'].lower():
-                                for call_idx, call_name in driver_calls:
-                                    approx_addr = hex(int(kernel_base, 16) if kernel_base else 0x1000000 + call_idx * 0x100)
-                                    entries.append({
-                                        'Index': call_idx,
-                                        'Address': approx_addr,
-                                        'Module': driver['Name'],
-                                        'Symbol': call_name,
-                                        'ServiceTable': 'KeServiceDescriptorTableShadow',
-                                        'Status': 'Normal'
-                                    })
-            
-            except Exception as e:
-                vollog.debug(f"Driver-based SSDT failed: {e}")
-                
-        except Exception as e:
-            vollog.debug(f"Error getting SSDT: {e}")
-            # Fallback to basic information
-            entries.append({
-                'Index': 0,
-                'Address': '0xfffff80000000000',
-                'Module': 'ntoskrnl.exe',
-                'Symbol': 'System Call Table',
-                'ServiceTable': 'KeServiceDescriptorTable',
-                'Status': 'Normal'
-            })
-        
-        # Ensure we have some entries even if all methods fail
-        if not entries:
-            entries.append({
-                'Index': 0,
-                'Address': '0xfffff80000000000',
-                'Module': 'ntoskrnl.exe',
-                'Symbol': 'NtCreateFile',
-                'ServiceTable': 'KeServiceDescriptorTable',
-                'Status': 'Normal'
-            })
-        
-        return entries
-    
-    @staticmethod
-    def get_kernel_timers():
-        """Get kernel timer information from system processes and threads."""
-        import datetime
-        timers = []
-        try:
-            current_time = datetime.datetime.now()
-            
-            for proc in psutil.process_iter():
-                try:
-                    pid = proc.pid
-                    proc_name = proc.name()
-                    proc_create_time = proc.create_time()
-                    
-                    uptime_seconds = (current_time.timestamp() - proc_create_time)
-                    
-                    try:
-                        threads = proc.threads()
-                        if threads:
-                            for idx, thread in enumerate(threads):
-                                total_time = thread.user_time + thread.system_time
-                                
-                                if total_time > 0:
-                                    next_fire = current_time + datetime.timedelta(seconds=total_time % 60)
-                                    due_time = next_fire.strftime('%Y-%m-%d %H:%M:%S')
-                                    signaled = 'Yes'
-                                else:
-                                    due_time = 'Pending'
-                                    signaled = 'No'
-                                
-                                vaddr, _ = LiveSystemContext.get_process_offsets(pid)
-                                if vaddr and vaddr > 0:
-                                    timer_offset = hex(vaddr + (idx * 0x1000))
-                                else:
-                                    timer_offset = hex(id(thread))
-                                
-                                timers.append({
-                                    'Offset': timer_offset,
-                                    'DueTime': due_time,
-                                    'Period': f"{total_time:.3f}s",
-                                    'Signaled': signaled,
-                                    'Module': proc_name[:24],
-                                    'Type': 'ThreadTimer'
-                                })
-                        else:
-                            vaddr, _ = LiveSystemContext.get_process_offsets(pid)
-                            timer_offset = hex(vaddr) if vaddr else hex(pid * 0x10000)
-                            
-                            created = datetime.datetime.fromtimestamp(proc_create_time)
-                            
-                            timers.append({
-                                'Offset': timer_offset,
-                                'DueTime': created.strftime('%Y-%m-%d %H:%M:%S'),
-                                'Period': f"{uptime_seconds:.1f}s",
-                                'Signaled': 'Yes',
-                                'Module': proc_name[:24],
-                                'Type': 'ProcessTimer'
-                            })
-                    except (psutil.AccessDenied, AttributeError, Exception):
-                        vaddr, _ = LiveSystemContext.get_process_offsets(pid)
-                        timer_offset = hex(vaddr) if vaddr else hex(pid * 0x10000)
-                        
-                        try:
-                            created = datetime.datetime.fromtimestamp(proc_create_time)
-                            due_time_str = created.strftime('%Y-%m-%d %H:%M:%S')
-                            period_str = f"{uptime_seconds:.1f}s"
-                        except:
-                            due_time_str = 'N/A'
-                            period_str = 'N/A'
-                        
-                        timers.append({
-                            'Offset': timer_offset,
-                            'DueTime': due_time_str,
-                            'Period': period_str,
-                            'Signaled': 'Yes',
-                            'Module': proc_name[:24],
-                            'Type': 'ProcessTimer'
-                        })
-                        
-                except (psutil.NoSuchProcess, psutil.AccessDenied, psutil.ZombieProcess, Exception):
-                    continue
-                    
-        except Exception as e:
-            vollog.debug(f"Error getting kernel timers: {e}")
-        
-        return timers
-    
-    @staticmethod
-    def get_console_history():
-        """Get command history from PowerShell and CMD."""
-        history = []
-        try:
-            ps_history_path = os.path.expandvars(r'%APPDATA%\Microsoft\Windows\PowerShell\PSReadLine\ConsoleHost_history.txt')
-            if os.path.exists(ps_history_path):
-                try:
-                    with open(ps_history_path, 'r', encoding='utf-8', errors='ignore') as f:
-                        for idx, line in enumerate(f):
-                            history.append({
-                                'Index': idx,
-                                'Application': 'PowerShell',
-                                'Command': line.strip(),
-                                'Source': 'PSReadLine'
-                            })
-                except Exception as e:
-                    vollog.debug(f"Error reading PS history: {e}")
-            
-            try:
-                import subprocess
-                result = subprocess.run(['doskey', '/history'], capture_output=True, text=True, shell=True)
-                if result.returncode == 0:
-                    for idx, line in enumerate(result.stdout.strip().split('\n')):
-                        if line.strip():
-                            history.append({
-                                'Index': idx,
-                                'Application': 'CMD',
-                                'Command': line.strip(),
-                                'Source': 'DOSKey'
-                            })
-            except Exception as e:
-                vollog.debug(f"Error getting CMD history: {e}")
-                
-        except Exception as e:
-            vollog.debug(f"Error getting console history: {e}")
-        
-        return history
-    
-    @staticmethod
-    def get_desktop_files():
-        """Get files and folders on the user's desktop."""
-        desktop_items = []
-        try:
-            user_desktop = os.path.expanduser("~\\Desktop")
-            public_desktop = os.path.expandvars(r"%PUBLIC%\Desktop")
-            
-            desktop_paths = [user_desktop]
-            if os.path.exists(public_desktop) and public_desktop != user_desktop:
-                desktop_paths.append(public_desktop)
-            
-            for desktop_path in desktop_paths:
-                if not os.path.exists(desktop_path):
-                    continue
-                
-                try:
-                    for item in os.listdir(desktop_path):
-                        item_path = os.path.join(desktop_path, item)
-                        try:
-                            stat_info = os.stat(item_path)
-                            is_dir = os.path.isdir(item_path)
-                            
-                            desktop_items.append({
-                                'Name': item,
-                                'Type': 'Folder' if is_dir else 'File',
-                                'Path': item_path,
-                                'Size': stat_info.st_size if not is_dir else 0,
-                                'Modified': stat_info.st_mtime,
-                                'Location': 'Public' if desktop_path == public_desktop else 'User'
-                            })
-                        except (OSError, PermissionError) as e:
-                            vollog.debug(f"Error accessing {item_path}: {e}")
-                            continue
-                except (OSError, PermissionError) as e:
-                    vollog.debug(f"Error listing desktop {desktop_path}: {e}")
-                    continue
-        except Exception as e:
-            vollog.debug(f"Error getting desktop files: {e}")
-        
-        return desktop_items
-    
-    @staticmethod
-    def get_desktops():
-        """Get desktop and window station information."""
-        desktops = []
-        try:
-            import win32gui
-            import win32process
-            
-            def enum_windows_callback(hwnd, results):
-                if win32gui.IsWindowVisible(hwnd):
-                    try:
-                        _, pid = win32process.GetWindowThreadProcessId(hwnd)
-                        title = win32gui.GetWindowText(hwnd)
-                        class_name = win32gui.GetClassName(hwnd)
-                        
-                        if title or class_name:
-                            try:
-                                proc = psutil.Process(pid)
-                                proc_name = proc.name()
-                            except:
-                                proc_name = 'Unknown'
-                            
-                            results.append({
-                                'Offset': hex(hwnd),
-                                'WindowStation': 'WinSta0',
-                                'Session': 1,
-                                'Desktop': 'Default',
-                                'Process': proc_name,
-                                'PID': pid,
-                                'Title': title[:50] if title else class_name[:50]
-                            })
-                    except Exception as e:
-                        vollog.debug(f"Error processing window: {e}")
-                return True
-            
-            results = []
-            win32gui.EnumWindows(enum_windows_callback, results)
-            desktops = results
-            
-        except Exception as e:
-            vollog.debug(f"Error enumerating desktops: {e}")
-            for proc in psutil.process_iter(['pid', 'name']):
-                try:
-                    if proc.info['name'] in ['explorer.exe', 'dwm.exe', 'csrss.exe']:
-                        desktops.append({
-                            'Offset': hex(proc.info['pid']),
-                            'WindowStation': 'WinSta0',
-                            'Session': 1,
-                            'Desktop': 'Default',
-                            'Process': proc.info['name'],
-                            'PID': proc.info['pid'],
-                            'Title': 'System Process'
-                        })
-                except (psutil.NoSuchProcess, psutil.AccessDenied):
-                    continue
-        
-        return desktops
-    
-    @staticmethod
-    def get_job_links():
-        """Get job object information for processes."""
-        job_links = []
-        try:
-            process_tree = defaultdict(list)
-            
-            for proc in psutil.process_iter(['pid', 'name', 'ppid', 'create_time']):
-                try:
-                    pinfo = proc.info
-                    pid = pinfo['pid']
-                    ppid = pinfo['ppid']
-                    process_tree[ppid].append(pid)
-                    
-                    children = process_tree.get(pid, [])
-                    if len(children) > 0:
-                        vaddr, _ = LiveSystemContext.get_process_offsets(pid)
-                        
-                        job_links.append({
-                            'Offset': hex(vaddr) if vaddr else hex(pid * 0x10000),
-                            'Name': pinfo['name'],
-                            'PID': pid,
-                            'PPID': ppid,
-                            'Session': 1,
-                            'JobSession': 1,
-                            'Wow64': False,
-                            'TotalProcs': len(children),
-                            'ActiveProcs': len(children),
-                            'TerminatedProcs': 0,
-                            'JobLink': 'Parent',
-                            'Path': 'N/A'
-                        })
-                        
-                        for child_pid in children:
-                            try:
-                                child_proc = psutil.Process(child_pid)
-                                child_vaddr, _ = LiveSystemContext.get_process_offsets(child_pid)
-                                
-                                job_links.append({
-                                    'Offset': hex(child_vaddr) if child_vaddr else hex(child_pid * 0x10000),
-                                    'Name': child_proc.name(),
-                                    'PID': child_pid,
-                                    'PPID': pid,
-                                    'Session': 1,
-                                    'JobSession': 1,
-                                    'Wow64': False,
-                                    'TotalProcs': 0,
-                                    'ActiveProcs': 0,
-                                    'TerminatedProcs': 0,
-                                    'JobLink': 'Child',
-                                    'Path': child_proc.exe() if child_proc.exe() else 'N/A'
-                                })
-                            except (psutil.NoSuchProcess, psutil.AccessDenied):
-                                continue
-                except (psutil.NoSuchProcess, psutil.AccessDenied):
-                    continue
-                    
-        except Exception as e:
-            vollog.debug(f"Error getting job links: {e}")
-        
-        return job_links
-    
-    @staticmethod
-    def get_scheduled_tasks_registry():
-        """Get scheduled tasks from Windows Task Scheduler."""
-        tasks = []
-        try:
-            import win32com.client
-            scheduler = win32com.client.Dispatch('Schedule.Service')
-            scheduler.Connect()
-            
-            def enum_tasks_recursive(folder, path='\\'):
-                try:
-                    task_collection = folder.GetTasks(0)
-                    for task in task_collection:
-                        try:
-                            definition = task.Definition
-                            
-                            triggers = []
-                            for trigger in definition.Triggers:
-                                trigger_type_map = {
-                                    0: 'Event', 1: 'Time', 2: 'Daily', 3: 'Weekly',
-                                    4: 'Monthly', 5: 'MonthlyDOW', 6: 'Idle',
-                                    7: 'Registration', 8: 'Boot', 9: 'Logon', 11: 'SessionStateChange'
-                                }
-                                trigger_type = trigger_type_map.get(trigger.Type, 'Unknown')
-                                triggers.append(trigger_type)
-                            
-                            actions = []
-                            for action in definition.Actions:
-                                if action.Type == 0:
-                                    actions.append(action.Path if hasattr(action, 'Path') else 'Unknown')
-                            
-                            state_map = {0: 'Unknown', 1: 'Disabled', 2: 'Queued', 3: 'Ready', 4: 'Running'}
-                            state = state_map.get(task.State, 'Unknown')
-                            
-                            tasks.append({
-                                'Name': task.Name,
-                                'Path': path + task.Name,
-                                'State': state,
-                                'Enabled': task.Enabled,
-                                'Hidden': definition.Settings.Hidden if hasattr(definition.Settings, 'Hidden') else False,
-                                'LastRun': str(task.LastRunTime) if task.LastRunTime else 'Never',
-                                'NextRun': str(task.NextRunTime) if task.NextRunTime else 'N/A',
-                                'Triggers': ', '.join(triggers) if triggers else 'None',
-                                'Actions': ', '.join(actions) if actions else 'None',
-                                'Author': definition.RegistrationInfo.Author if hasattr(definition.RegistrationInfo, 'Author') else 'N/A'
-                            })
-                        except Exception as e:
-                            vollog.debug(f"Error processing task {task.Name}: {e}")
-                    
-                    subfolders = folder.GetFolders(0)
-                    for subfolder in subfolders:
-                        enum_tasks_recursive(subfolder, path + subfolder.Name + '\\')
-                        
-                except Exception as e:
-                    vollog.debug(f"Error enumerating folder: {e}")
-            
-            root_folder = scheduler.GetFolder('\\')
-            enum_tasks_recursive(root_folder)
-            
-        except Exception as e:
-            vollog.debug(f"Error getting scheduled tasks: {e}")
-        
-        return tasks
-    
-    @staticmethod
-    def get_deleted_files():
-        """Get recently deleted files from Recycle Bin and registry traces."""
-        deleted_files = []
-        
-        # Scan Recycle Bin using Shell API
-        try:
-            import win32com.client
-            shell = win32com.client.Dispatch("Shell.Application")
-            
-            recycler = shell.NameSpace(10)
-            if recycler:
-                for item in recycler.Items():
-                    try:
-                        deleted_files.append({
-                            'OriginalName': item.Name,
-                            'DeletedPath': item.Path,
-                            'Size': item.Size,
-                            'DateDeleted': str(item.ModifyDate) if hasattr(item, 'ModifyDate') else 'N/A',
-                            'Type': item.Type
-                        })
-                    except Exception as e:
-                        vollog.debug(f"Error reading recycle item: {e}")
-        except Exception as e:
-            vollog.debug(f"Error accessing Recycle Bin via Shell: {e}")
-        
-        # Scan filesystem Recycle Bin
-        user_sid_pattern = os.path.expandvars(r'%SystemDrive%\$Recycle.Bin')
-        if os.path.exists(user_sid_pattern):
-            for sid_folder in os.listdir(user_sid_pattern):
-                recycle_path = os.path.join(user_sid_pattern, sid_folder)
-                if os.path.isdir(recycle_path):
-                    try:
-                        for item in os.listdir(recycle_path):
-                            item_path = os.path.join(recycle_path, item)
-                            try:
-                                stat_info = os.stat(item_path)
-                                import datetime
-                                deleted_files.append({
-                                    'OriginalName': item,
-                                    'DeletedPath': item_path,
-                                    'Size': stat_info.st_size,
-                                    'DateDeleted': datetime.datetime.fromtimestamp(stat_info.st_mtime).strftime('%Y-%m-%d %H:%M:%S'),
-                                    'Type': os.path.splitext(item)[1] or 'Unknown'
-                                })
-                            except Exception as e:
-                                vollog.debug(f"Error reading {item_path}: {e}")
-                    except Exception as e:
-                        vollog.debug(f"Error scanning {recycle_path}: {e}")
-        
-        # Scan registry for deleted file traces
-        try:
-            # Check BagMRU (tracks deleted folders/files)
-            key_path = r"Software\Microsoft\Windows\Shell\BagMRU"
-            try:
-                key = winreg.OpenKey(winreg.HKEY_CURRENT_USER, key_path, 0, winreg.KEY_READ)
-                i = 0
-                while True:
-                    try:
-                        subkey_name = winreg.EnumKey(key, i)
-                        subkey = winreg.OpenKey(key, subkey_name, 0, winreg.KEY_READ)
-                        try:
-                            target_path, _ = winreg.QueryValueEx(subkey, "Target")
-                            deleted_files.append({
-                                'OriginalName': os.path.basename(target_path) if target_path else 'Unknown',
-                                'DeletedPath': f"Registry:BagMRU\\{subkey_name}",
-                                'Size': 0,
-                                'DateDeleted': 'N/A',
-                                'Type': os.path.splitext(target_path)[1] if target_path else 'Folder'
-                            })
-                        except:
-                            pass
-                        winreg.CloseKey(subkey)
-                        i += 1
-                    except OSError:
-                        break
-                winreg.CloseKey(key)
-            except Exception as e:
-                vollog.debug(f"Error reading BagMRU: {e}")
-            
-            # Check WordWheelQuery (search history, may contain deleted file names)
-            key_path = r"Software\Microsoft\Windows\CurrentVersion\Explorer\WordWheelQuery"
-            try:
-                key = winreg.OpenKey(winreg.HKEY_CURRENT_USER, key_path, 0, winreg.KEY_READ)
-                i = 0
-                while True:
-                    try:
-                        value_name, value_data, value_type = winreg.EnumValue(key, i)
-                        if isinstance(value_data, bytes):
-                            try:
-                                search_term = value_data.decode('utf-16-le', errors='ignore').rstrip('\x00')
-                                if search_term and len(search_term) > 0:
-                                    # Check if it looks like a filename
-                                    if '.' in search_term and not search_term.startswith('http'):
-                                        deleted_files.append({
-                                            'OriginalName': search_term,
-                                            'DeletedPath': 'Registry:SearchHistory',
-                                            'Size': 0,
-                                            'DateDeleted': 'N/A',
-                                            'Type': os.path.splitext(search_term)[1] if '.' in search_term else 'Unknown'
-                                        })
-                            except:
-                                pass
-                        i += 1
-                    except OSError:
-                        break
-                winreg.CloseKey(key)
-            except Exception as e:
-                vollog.debug(f"Error reading WordWheelQuery: {e}")
-            
-            # Check TypedPaths (deleted paths from Explorer address bar)
-            key_path = r"Software\Microsoft\Windows\CurrentVersion\Explorer\TypedPaths"
-            try:
-                key = winreg.OpenKey(winreg.HKEY_CURRENT_USER, key_path, 0, winreg.KEY_READ)
-                i = 0
-                while True:
-                    try:
-                        value_name, value_data, value_type = winreg.EnumValue(key, i)
-                        if value_data and isinstance(value_data, str):
-                            # Check if path exists, if not it might be deleted
-                            if not os.path.exists(value_data):
-                                deleted_files.append({
-                                    'OriginalName': os.path.basename(value_data),
-                                    'DeletedPath': f"Registry:TypedPaths -> {value_data}",
-                                    'Size': 0,
-                                    'DateDeleted': 'N/A',
-                                    'Type': os.path.splitext(value_data)[1] if '.' in value_data else 'Folder'
-                                })
-                        i += 1
-                    except OSError:
-                        break
-                winreg.CloseKey(key)
-            except Exception as e:
-                vollog.debug(f"Error reading TypedPaths: {e}")
-                
-        except Exception as e:
-            vollog.debug(f"Error scanning registry for deleted files: {e}")
-        
-        return deleted_files
-    
-    @staticmethod
-    def get_recent_documents():
-        """Get recently opened documents from Windows Recent folder, registry, and Office MRU."""
-        recent_docs = []
-        
-        # Scan Recent folder (shortcuts)
-        recent_folder = os.path.expandvars(r'%APPDATA%\Microsoft\Windows\Recent')
-        
-        if os.path.exists(recent_folder):
-            try:
-                for item in os.listdir(recent_folder):
-                    if item.endswith('.lnk'):
-                        link_path = os.path.join(recent_folder, item)
-                        try:
-                            import win32com.client
-                            shell = win32com.client.Dispatch("WScript.Shell")
-                            shortcut = shell.CreateShortCut(link_path)
-                            target_path = shortcut.Targetpath
-                            
-                            stat_info = os.stat(link_path)
-                            import datetime
-                            
-                            recent_docs.append({
-                                'DocumentName': item[:-4],
-                                'TargetPath': target_path,
-                                'LinkPath': link_path,
-                                'LastAccessed': datetime.datetime.fromtimestamp(stat_info.st_atime).strftime('%Y-%m-%d %H:%M:%S'),
-                                'Modified': datetime.datetime.fromtimestamp(stat_info.st_mtime).strftime('%Y-%m-%d %H:%M:%S'),
-                                'Type': os.path.splitext(target_path)[1] if target_path else 'Unknown'
-                            })
-                        except Exception as e:
-                            vollog.debug(f"Error reading shortcut {link_path}: {e}")
-            except Exception as e:
-                vollog.debug(f"Error scanning recent folder: {e}")
-        
-        # Scan RecentDocs registry
-        try:
-            key_path = r"Software\Microsoft\Windows\CurrentVersion\Explorer\RecentDocs"
-            key = winreg.OpenKey(winreg.HKEY_CURRENT_USER, key_path, 0, winreg.KEY_READ)
-            
-            i = 0
-            while True:
-                try:
-                    value_name, value_data, value_type = winreg.EnumValue(key, i)
-                    if isinstance(value_data, bytes):
-                        try:
-                            doc_name = value_data.decode('utf-16-le', errors='ignore').rstrip('\x00')
-                            if doc_name and len(doc_name) > 0:
-                                recent_docs.append({
-                                    'DocumentName': doc_name,
-                                    'TargetPath': 'N/A',
-                                    'LinkPath': 'Registry:RecentDocs',
-                                    'LastAccessed': 'N/A',
-                                    'Modified': 'N/A',
-                                    'Type': os.path.splitext(doc_name)[1] if '.' in doc_name else 'Unknown'
-                                })
-                        except:
-                            pass
-                    i += 1
-                except OSError:
-                    break
-            
-            winreg.CloseKey(key)
-        except Exception as e:
-            vollog.debug(f"Error reading registry recent docs: {e}")
-        
-        # Scan Office MRU (Most Recently Used)
-        office_versions = ['16.0', '15.0', '14.0', '12.0', '11.0']  # Office 2016-2003
-        office_apps = ['Word', 'Excel', 'PowerPoint', 'Access']
-        
-        for version in office_versions:
-            for app in office_apps:
-                try:
-                    key_path = rf"Software\Microsoft\Office\{version}\{app}\File MRU"
-                    try:
-                        key = winreg.OpenKey(winreg.HKEY_CURRENT_USER, key_path, 0, winreg.KEY_READ)
-                        i = 0
-                        while True:
-                            try:
-                                value_name, value_data, value_type = winreg.EnumValue(key, i)
-                                if value_name.startswith('Item'):
-                                    # Parse MRU format: [F00000000][T01D9...][O00000000]*C:\path\to\file.docx
-                                    if isinstance(value_data, str) and '*' in value_data:
-                                        file_path = value_data.split('*')[-1]
-                                        recent_docs.append({
-                                            'DocumentName': os.path.basename(file_path),
-                                            'TargetPath': file_path,
-                                            'LinkPath': f'Registry:Office_{app}_MRU',
-                                            'LastAccessed': 'N/A',
-                                            'Modified': 'N/A',
-                                            'Type': os.path.splitext(file_path)[1] if '.' in file_path else 'Unknown'
-                                        })
-                                i += 1
-                            except OSError:
-                                break
-                        winreg.CloseKey(key)
-                    except FileNotFoundError:
-                        continue
-                except Exception as e:
-                    vollog.debug(f"Error reading Office {app} MRU: {e}")
-        
-        # Scan UserAssist (tracks program execution, including document opens)
-        try:
-            key_path = r"Software\Microsoft\Windows\CurrentVersion\Explorer\UserAssist"
-            key = winreg.OpenKey(winreg.HKEY_CURRENT_USER, key_path, 0, winreg.KEY_READ)
-            
-            i = 0
-            while True:
-                try:
-                    guid_key_name = winreg.EnumKey(key, i)
-                    guid_key = winreg.OpenKey(key, guid_key_name + r"\Count", 0, winreg.KEY_READ)
-                    
-                    j = 0
-                    while True:
-                        try:
-                            value_name, value_data, value_type = winreg.EnumValue(guid_key, j)
-                            # ROT13 decode the value name
-                            import codecs
-                            decoded_name = codecs.decode(value_name, 'rot13')
-                            
-                            # Check if it's a document
-                            doc_extensions = ['.doc', '.docx', '.xls', '.xlsx', '.ppt', '.pptx', '.pdf', '.txt']
-                            if any(decoded_name.lower().endswith(ext) for ext in doc_extensions):
-                                recent_docs.append({
-                                    'DocumentName': os.path.basename(decoded_name),
-                                    'TargetPath': decoded_name,
-                                    'LinkPath': 'Registry:UserAssist',
-                                    'LastAccessed': 'N/A',
-                                    'Modified': 'N/A',
-                                    'Type': os.path.splitext(decoded_name)[1]
-                                })
-                            j += 1
-                        except OSError:
-                            break
-                    
-                    winreg.CloseKey(guid_key)
-                    i += 1
-                except OSError:
-                    break
-            
-            winreg.CloseKey(key)
-        except Exception as e:
-            vollog.debug(f"Error reading UserAssist: {e}")
-        
-        return recent_docs
-    
-    @staticmethod
-    def get_userassist():
-        """Get UserAssist data - tracks program execution with timestamps and run counts."""
-        userassist_data = []
-        
-        try:
-            key_path = r"Software\Microsoft\Windows\CurrentVersion\Explorer\UserAssist"
-            key = winreg.OpenKey(winreg.HKEY_CURRENT_USER, key_path, 0, winreg.KEY_READ)
-            
-            i = 0
-            while True:
-                try:
-                    guid_key_name = winreg.EnumKey(key, i)
-                    guid_key = winreg.OpenKey(key, guid_key_name + r"\Count", 0, winreg.KEY_READ)
-                    
-                    j = 0
-                    while True:
-                        try:
-                            value_name, value_data, value_type = winreg.EnumValue(guid_key, j)
-                            
-                            # ROT13 decode the value name
-                            import codecs
-                            decoded_name = codecs.decode(value_name, 'rot13')
-                            
-                            # Parse the binary data for run count and timestamp
-                            run_count = 0
-                            last_executed = 'N/A'
-                            
-                            if len(value_data) >= 16:
-                                try:
-                                    # Bytes 4-7: Run count
-                                    run_count = struct.unpack('<I', value_data[4:8])[0]
-                                    
-                                    # Bytes 60-67: FILETIME timestamp
-                                    if len(value_data) >= 68:
-                                        timestamp_low = struct.unpack('<I', value_data[60:64])[0]
-                                        timestamp_high = struct.unpack('<I', value_data[64:68])[0]
-                                        timestamp = (timestamp_high << 32) | timestamp_low
-                                        
-                                        if timestamp > 0:
-                                            import datetime
-                                            # Convert Windows FILETIME to datetime
-                                            EPOCH_AS_FILETIME = 116444736000000000
-                                            timestamp_seconds = (timestamp - EPOCH_AS_FILETIME) / 10000000
-                                            last_executed = datetime.datetime.fromtimestamp(timestamp_seconds).strftime('%Y-%m-%d %H:%M:%S')
-                                except:
-                                    pass
-                            
-                            userassist_data.append({
-                                'Program': decoded_name,
-                                'RunCount': run_count if run_count > 0 else 1,
-                                'LastExecuted': last_executed,
-                                'GUID': guid_key_name,
-                                'Type': 'Application' if decoded_name.endswith('.exe') else 'Shortcut/File'
-                            })
-                            
-                            j += 1
-                        except OSError:
-                            break
-                    
-                    winreg.CloseKey(guid_key)
-                    i += 1
-                except OSError:
-                    break
-            
-            winreg.CloseKey(key)
-        except Exception as e:
-            vollog.debug(f"Error reading UserAssist: {e}")
-        
-        return userassist_data
-    
-    @staticmethod
-    def get_persistence_mechanisms():
-        """Get persistence mechanisms - startup programs, scheduled tasks, and suspicious executables."""
-        persistence_items = []
-        
-        # 1. Startup folders
-        startup_locations = [
-            (os.path.expandvars(r'%APPDATA%\Microsoft\Windows\Start Menu\Programs\Startup'), 'User Startup'),
-            (os.path.expandvars(r'%ProgramData%\Microsoft\Windows\Start Menu\Programs\Startup'), 'All Users Startup'),
-        ]
-        
-        for startup_path, location_name in startup_locations:
-            if os.path.exists(startup_path):
-                try:
-                    for item in os.listdir(startup_path):
-                        item_path = os.path.join(startup_path, item)
-                        try:
-                            stat_info = os.stat(item_path)
-                            import datetime
-                            
-                            target = item_path
-                            # If it's a shortcut, get the target
-                            if item.endswith('.lnk'):
-                                try:
-                                    import win32com.client
-                                    shell = win32com.client.Dispatch("WScript.Shell")
-                                    shortcut = shell.CreateShortCut(item_path)
-                                    target = shortcut.Targetpath
-                                except:
-                                    pass
-                            
-                            persistence_items.append({
-                                'Name': item,
-                                'Location': location_name,
-                                'Path': item_path,
-                                'Target': target,
-                                'Created': datetime.datetime.fromtimestamp(stat_info.st_ctime).strftime('%Y-%m-%d %H:%M:%S'),
-                                'Type': 'Startup Folder'
-                            })
-                        except Exception as e:
-                            vollog.debug(f"Error reading {item_path}: {e}")
-                except Exception as e:
-                    vollog.debug(f"Error scanning {startup_path}: {e}")
-        
-        # 2. Registry Run keys
-        run_keys = [
-            (winreg.HKEY_CURRENT_USER, r"Software\Microsoft\Windows\CurrentVersion\Run", "HKCU Run"),
-            (winreg.HKEY_CURRENT_USER, r"Software\Microsoft\Windows\CurrentVersion\RunOnce", "HKCU RunOnce"),
-            (winreg.HKEY_LOCAL_MACHINE, r"Software\Microsoft\Windows\CurrentVersion\Run", "HKLM Run"),
-            (winreg.HKEY_LOCAL_MACHINE, r"Software\Microsoft\Windows\CurrentVersion\RunOnce", "HKLM RunOnce"),
-            (winreg.HKEY_CURRENT_USER, r"Software\Microsoft\Windows\CurrentVersion\Explorer\User Shell Folders", "User Shell Folders"),
-            (winreg.HKEY_CURRENT_USER, r"Software\Microsoft\Windows\CurrentVersion\Explorer\Shell Folders", "Shell Folders"),
-        ]
-        
-        for hive, key_path, location_name in run_keys:
-            try:
-                key = winreg.OpenKey(hive, key_path, 0, winreg.KEY_READ)
-                i = 0
-                while True:
-                    try:
-                        value_name, value_data, value_type = winreg.EnumValue(key, i)
-                        if value_data:
-                            persistence_items.append({
-                                'Name': value_name,
-                                'Location': location_name,
-                                'Path': key_path,
-                                'Target': str(value_data),
-                                'Created': 'N/A',
-                                'Type': 'Registry Run Key'
-                            })
-                        i += 1
-                    except OSError:
-                        break
-                winreg.CloseKey(key)
-            except Exception as e:
-                vollog.debug(f"Error reading {key_path}: {e}")
-        
-        # 3. Scan for .exe files in suspicious locations
-        suspicious_dirs = [
-            (os.path.expandvars(r'%TEMP%'), 'TEMP'),
-            (os.path.expandvars(r'%APPDATA%'), 'APPDATA'),
-            (os.path.expandvars(r'%LOCALAPPDATA%'), 'LOCALAPPDATA'),
-        ]
-        
-        for base_dir, dir_name in suspicious_dirs:
-            if os.path.exists(base_dir):
-                try:
-                    # Scan up to 3 levels deep
-                    for root, dirs, files in os.walk(base_dir):
-                        # Limit depth
-                        depth = root[len(base_dir):].count(os.sep)
-                        if depth > 2:
-                            dirs.clear()
-                            continue
-                        
-                        for file in files:
-                            if file.lower().endswith('.exe'):
-                                file_path = os.path.join(root, file)
-                                try:
-                                    stat_info = os.stat(file_path)
-                                    import datetime
-                                    
-                                    persistence_items.append({
-                                        'Name': file,
-                                        'Location': f'{dir_name} ({os.path.dirname(file_path)})',
-                                        'Path': file_path,
-                                        'Target': file_path,
-                                        'Created': datetime.datetime.fromtimestamp(stat_info.st_ctime).strftime('%Y-%m-%d %H:%M:%S'),
-                                        'Type': 'Suspicious EXE'
-                                    })
-                                except Exception as e:
-                                    vollog.debug(f"Error reading {file_path}: {e}")
-                except Exception as e:
-                    vollog.debug(f"Error scanning {base_dir}: {e}")
-        
-        # 4. Scheduled Tasks (already have method, just reference)
-        try:
-            import win32com.client
-            scheduler = win32com.client.Dispatch('Schedule.Service')
-            scheduler.Connect()
-            
-            def enum_tasks_recursive(folder, path='\\'):
-                try:
-                    task_collection = folder.GetTasks(0)
-                    for task in task_collection:
-                        try:
-                            definition = task.Definition
-                            actions = []
-                            for action in definition.Actions:
-                                if action.Type == 0:
-                                    actions.append(action.Path if hasattr(action, 'Path') else 'Unknown')
-                            
-                            if actions:
-                                persistence_items.append({
-                                    'Name': task.Name,
-                                    'Location': 'Task Scheduler',
-                                    'Path': path + task.Name,
-                                    'Target': ', '.join(actions),
-                                    'Created': str(definition.RegistrationInfo.Date) if hasattr(definition.RegistrationInfo, 'Date') else 'N/A',
-                                    'Type': 'Scheduled Task'
-                                })
-                        except:
-                            pass
-                    
-                    subfolders = folder.GetFolders(0)
-                    for subfolder in subfolders:
-                        enum_tasks_recursive(subfolder, path + subfolder.Name + '\\')
-                except:
-                    pass
-            
-            root_folder = scheduler.GetFolder('\\')
-            enum_tasks_recursive(root_folder)
-        except Exception as e:
-            vollog.debug(f"Error getting scheduled tasks: {e}")
-        
-        return persistence_items
-    
-    @staticmethod
-    def generate_process_graph_data():
-        """Generate process tree graph data for proccon visualization."""
-        processes = LiveSystemContext.get_processes()
-        
-        # Build process dictionary
-        proc_dict = {}
-        for proc in processes:
-            pid = proc['PID']
-            proc_dict[pid] = {
-                'name': proc['Name'],
-                'pid': pid,
-                'ppid': proc['PPID'],
-                'path': proc['Path'],
-                'threads': proc['Threads'],
-                'children': []
-            }
-        
-        # Build parent-child relationships
-        for pid, proc_data in proc_dict.items():
-            ppid = proc_data['ppid']
-            if ppid in proc_dict:
-                proc_dict[ppid]['children'].append(pid)
-        
-        return proc_dict
-    
-    @staticmethod
-    def generate_dot_file(output_path: str = "process_tree.dot"):
-        """Generate Graphviz DOT file for process tree visualization."""
-        processes = LiveSystemContext.generate_process_graph_data()
-        
-        dot_lines = []
-        
-        # Header
-        dot_lines.append("digraph ProcessTree {")
-        dot_lines.append("    rankdir=TB;")
-        dot_lines.append("    node [shape=box, style=filled];")
-        dot_lines.append("    edge [fontsize=10];")
-        dot_lines.append("    ")
-        
-        # Find root processes
-        root_pids = []
-        for pid, proc_data in processes.items():
-            ppid = proc_data['ppid']
-            if ppid == 0 or ppid not in processes:
-                root_pids.append(pid)
-        
-        # Calculate process depth
-        def calculate_depth(pid, visited=None):
-            if visited is None:
-                visited = set()
-            if pid in visited:
-                return 0
-            visited.add(pid)
-            ppid = processes[pid]['ppid']
-            if ppid not in processes:
-                return 0
-            return 1 + calculate_depth(ppid, visited)
-        
-        # Add process nodes
-        for pid, proc_data in processes.items():
-            name = proc_data['name']
-            path = proc_data['path']
-            threads = proc_data['threads']
-            ppid = proc_data['ppid']
-            children = proc_data['children']
-            
-            # Determine role
-            depth = calculate_depth(pid)
-            if pid in root_pids:
-                role = "ROOT"
-                fillcolor = "lightcoral"
-            elif children:
-                role = "PARENT" if depth == 1 else f"PARENT (L{depth})"
-                fillcolor = "lightgreen"
-            else:
-                role = "CHILD" if depth == 1 else f"CHILD (L{depth})"
-                fillcolor = "lightblue"
-            
-            # Escape special characters
-            safe_name = name.replace('"', '\\"')
-            safe_path = path.replace('"', '\\"').replace('\\', '\\\\')
-            
-            # Build label
-            label = f"[{role}]\\n{safe_name}\\nPID: {pid} | PPID: {ppid}"
-            
-            if safe_path != safe_name and safe_path != "N/A":
-                if '\\\\' in safe_path:
-                    parts = safe_path.split('\\\\')
-                    if len(parts) > 3:
-                        safe_path = parts[0] + '\\\\...\\\\' + '\\\\'.join(parts[-2:])
-                label += f"\\nPath: {safe_path}"
-            
-            if threads > 0:
-                label += f"\\nThreads: {threads}"
-            
-            if children:
-                label += f"\\nChildren: {len(children)}"
-            
-            dot_lines.append(
-                f'    proc_{pid} [label="{label}", fillcolor="{fillcolor}"];'
-            )
-        
-        dot_lines.append("    ")
-        dot_lines.append("    // Parent-child relationships")
-        
-        # Add edges
-        edge_count = 0
-        for pid, proc_data in processes.items():
-            ppid = proc_data['ppid']
-            if ppid in processes:
-                child_name = proc_data['name']
-                edge_label = f"spawned\\n{child_name}"
-                
-                dot_lines.append(
-                    f'    proc_{ppid} -> proc_{pid} [label="{edge_label}", color="darkblue", penwidth=2];'
-                )
-                edge_count += 1
-        
-        dot_lines.append("    ")
-        dot_lines.append("    // Legend")
-        dot_lines.append('    subgraph cluster_legend {')
-        dot_lines.append('        label="Legend";')
-        dot_lines.append('        style=filled;')
-        dot_lines.append('        color=lightgrey;')
-        dot_lines.append('        node [shape=box, fontsize=10];')
-        dot_lines.append('        legend_root [label="ROOT\\n(No parent)", fillcolor="lightcoral"];')
-        dot_lines.append('        legend_parent [label="PARENT\\n(Has children)", fillcolor="lightgreen"];')
-        dot_lines.append('        legend_child [label="CHILD\\n(Leaf process)", fillcolor="lightblue"];')
-        dot_lines.append('        legend_root -> legend_parent [label="spawned", fontsize=9];')
-        dot_lines.append('        legend_parent -> legend_child [label="spawned", fontsize=9];')
-        dot_lines.append('    }')
-        dot_lines.append("}")
-        
-        # Write to file
-        try:
-            with open(output_path, 'w', encoding='utf-8') as f:
-                f.write("\n".join(dot_lines))
-            return True, len(processes), edge_count, len(root_pids)
-        except Exception as e:
-            vollog.error(f"Error writing DOT file: {e}")
-            return False, 0, 0, 0
-
-
-class FilelessMalwareHunter:
-    """Advanced fileless malware and in-memory threat detection for live systems."""
-    
-    # Enhanced PowerShell patterns
-    POWERSHELL_PATTERNS = {
-        'encoded_command': (rb'-e[ncodedcommand]*\s+[A-Za-z0-9+/=]{50,}', 'Critical'),
-        'invoke_expression': (rb'(?i)(iex|invoke-expression)', 'High'),
-        'download_string': (rb'(?i)(downloadstring|downloadfile|downloaddata)', 'Critical'),
-        'invoke_webrequest': (rb'(?i)(invoke-webrequest|invoke-restmethod)', 'High'),
-        'net_webclient': (rb'(?i)system\.net\.webclient', 'High'),
-        'hidden_window': (rb'(?i)-windowstyle\s+hidden', 'High'),
-        'bypass_execution': (rb'(?i)-executionpolicy\s+bypass', 'High'),
-        'encoded_script': (rb'(?i)frombase64string', 'High'),
-        'reflection_assembly': (rb'(?i)\[reflection\.assembly\]::load', 'Critical'),
-        'invoke_mimikatz': (rb'(?i)invoke-mimikatz', 'Critical'),
-        'invoke_shellcode': (rb'(?i)invoke-shellcode', 'Critical'),
-        'compressed_script': (rb'(?i)io\.compression\.gzipstream', 'Medium'),
-        'obfuscation': (rb'(?i)(\^|\`|"|\+){10,}', 'High'),
-        'memory_stream': (rb'(?i)system\.io\.memorystream', 'Medium'),
-        'process_start': (rb'(?i)system\.diagnostics\.process]::start', 'High'),
-        'amsi_bypass': (rb'(?i)(amsiutils|amsiinitfailed|amsi\.dll)', 'Critical'),
-        'empire_framework': (rb'(?i)(invoke-empire|get-empire)', 'Critical'),
-        'cobalt_strike': (rb'(?i)(invoke-beacon|invoke-dllinjection)', 'Critical'),
-        'metasploit': (rb'(?i)(invoke-meterpreter|invoke-payload)', 'Critical'),
-        'psattack': (rb'(?i)psattack', 'Critical'),
-        'nishang': (rb'(?i)invoke-powershelltcp', 'Critical'),
-    }
-    
-    # Credential theft patterns
-    CREDENTIAL_PATTERNS = {
-        'mimikatz_signature': (rb'(?i)(gentilkiwi|benjamin delpy|sekurlsa|lsadump)', 'Critical'),
-        'lsass_dump': (rb'(?i)(lsass\.exe|lsass\.dmp|procdump.*lsass)', 'Critical'),
-        'sam_dump': (rb'(?i)(reg.*save.*sam|reg.*save.*system)', 'Critical'),
-        'credential_manager': (rb'(?i)vaultcmd', 'High'),
-        'ntds_dit': (rb'(?i)ntds\.dit', 'Critical'),
-        'cached_credentials': (rb'(?i)mscash', 'High'),
-        'kerberos_ticket': (rb'(?i)(kirbi|invoke-kerberoast)', 'Critical'),
-        'dpapi': (rb'(?i)dpapi', 'High'),
-        'laZagne': (rb'(?i)lazagne', 'Critical'),
-        'password_filter': (rb'(?i)passwordchangenotify', 'Critical'),
-    }
-    
-    # WMI attack patterns
-    WMI_PATTERNS = {
-        'wmi_process_create': (rb'(?i)win32_process.*create', 'High'),
-        'wmi_event_consumer': (rb'(?i)commandlineeventconsumer', 'Critical'),
-        'wmi_persistence': (rb'(?i)__eventfilter', 'Critical'),
-        'wmi_lateral_movement': (rb'(?i)win32_scheduledjob', 'High'),
-        'wmi_exec': (rb'(?i)wmiexec', 'High'),
-    }
-    
-    # Persistence mechanisms
-    PERSISTENCE_PATTERNS = {
-        'registry_run': (rb'(?i)(software\\microsoft\\windows\\currentversion\\run)', 'High'),
-        'startup_folder': (rb'(?i)(startup|start menu\\programs\\startup)', 'High'),
-        'scheduled_task': (rb'(?i)(schtasks.*create|register-scheduledtask)', 'High'),
-        'service_creation': (rb'(?i)(sc.*create|new-service)', 'High'),
-        'dll_hijacking': (rb'(?i)(dll.*order|dll.*search)', 'Medium'),
-        'winlogon': (rb'(?i)software\\microsoft\\windows nt\\currentversion\\winlogon', 'High'),
-        'image_file_execution': (rb'(?i)image file execution options', 'High'),
-        'app_init': (rb'(?i)appinit_dlls', 'High'),
-        'bits_jobs': (rb'(?i)start-bitstransfer', 'Medium'),
-    }
-    
-    # Lateral movement patterns
-    LATERAL_MOVEMENT_PATTERNS = {
-        'psexec': (rb'(?i)psexe(c|csvc)', 'Critical'),
-        'remote_exec': (rb'(?i)(winrs|wmic.*node)', 'High'),
-        'smb_exec': (rb'(?i)(\\\\.*\\admin\$|\\\\.*\\c\$|\\\\.*\\ipc\$)', 'High'),
-        'rdp_usage': (rb'(?i)(mstsc|terminal.*server)', 'Medium'),
-        'pass_the_hash': (rb'(?i)(sekurlsa::pth|invoke-pth)', 'Critical'),
-        'pass_the_ticket': (rb'(?i)(kerberos::ptt|invoke-mimikatz.*ticket)', 'Critical'),
-        'dcom_exec': (rb'(?i)mmc20\.application', 'High'),
-        'wmi_exec_remote': (rb'(?i)invoke-wmimethod.*-computer', 'High'),
-    }
-    
-    # Defense evasion patterns  
-    EVASION_PATTERNS = {
-        'disable_defender': (rb'(?i)(set-mppreference.*-disable|add-mppreference.*-exclusion)', 'Critical'),
-        'clear_logs': (rb'(?i)(wevtutil.*cl|clear-eventlog)', 'Critical'),
-        'timestomp': (rb'(?i)timestomp', 'High'),
-        'sandbox_detection': (rb'(?i)(check.*vm|detect.*sandbox)', 'Medium'),
-        'disable_firewall': (rb'(?i)(netsh.*firewall.*off|set-netfirewallprofile)', 'High'),
-        'disable_uac': (rb'(?i)enablelua.*0', 'High'),
-        'process_masquerading': (rb'(?i)(svchost\.exe.*-k|rundll32.*,#1)', 'High'),
-        'rootkit': (rb'(?i)(zwsetsysteminformation|ntquerysysteminformation)', 'Critical'),
-    }
-    
-    # Malware families signatures
-    MALWARE_FAMILIES = {
-        'emotet': (rb'(?i)(emotet|epoch[1-5])', 'Critical'),
-        'trickbot': (rb'(?i)trickbot', 'Critical'),
-        'ryuk': (rb'(?i)(ryuk|hermes)', 'Critical'),
-        'conti': (rb'(?i)conti', 'Critical'),
-        'lockbit': (rb'(?i)lockbit', 'Critical'),
-        'qakbot': (rb'(?i)(qakbot|qbot)', 'Critical'),
-        'dridex': (rb'(?i)dridex', 'Critical'),
-        'icedid': (rb'(?i)icedid', 'Critical'),
-        'bumblebee': (rb'(?i)bumblebee', 'Critical'),
-    }
-    
-    # Cryptocurrency Miners signatures
-    CRYPTO_MINER_SIGNATURES = {
-        # Popular Mining Software
-        'xmrig': (rb'(?i)(xmrig|monero.*miner)', 'Critical'),
-        'xmrig_config': (rb'(?i)(donate-level|algo.*randomx|pool.*xmr)', 'Critical'),
-        'claymore': (rb'(?i)(claymore|ethminer)', 'Critical'),
-        'phoenixminer': (rb'(?i)phoenixminer', 'Critical'),
-        'cgminer': (rb'(?i)cgminer', 'Critical'),
-        'bfgminer': (rb'(?i)bfgminer', 'Critical'),
-        'nicehash': (rb'(?i)(nicehash|nhminer)', 'Critical'),
-        'minergate': (rb'(?i)minergate', 'Critical'),
-        'ccminer': (rb'(?i)ccminer', 'Critical'),
-        'cpuminer': (rb'(?i)(cpuminer|minerd)', 'Critical'),
-        'ethminer': (rb'(?i)ethminer', 'Critical'),
-        'ewbf': (rb'(?i)ewbf', 'Critical'),
-        'bminer': (rb'(?i)bminer', 'Critical'),
-        'gminer': (rb'(?i)gminer', 'Critical'),
-        'lolminer': (rb'(?i)lolminer', 'Critical'),
-        'nbminer': (rb'(?i)nbminer', 'Critical'),
-        'trex_miner': (rb'(?i)(t-rex|trex.*miner)', 'Critical'),
-        'teamredminer': (rb'(?i)teamredminer', 'Critical'),
-        
-        # Mining Pools
-        'mining_pool_stratum': (rb'(?i)(stratum\+tcp|stratum\+ssl)', 'High'),
-        'monero_pool': (rb'(?i)(xmr.*pool|moneroocean|supportxmr|minexmr)', 'High'),
-        'ethereum_pool': (rb'(?i)(ethermine|nanopool|f2pool.*eth)', 'High'),
-        'bitcoin_pool': (rb'(?i)(slushpool|antpool|btc\.com)', 'High'),
-        'mining_pool_url': (rb'(?i)(pool\..*\..*:3333|pool\..*\..*:4444|pool\..*\..*:5555)', 'High'),
-        
-        # Mining Algorithms
-        'cryptonight': (rb'(?i)(cryptonight|cryptonight-lite|cn/r)', 'High'),
-        'randomx': (rb'(?i)(randomx|rx/0)', 'High'),
-        'ethash': (rb'(?i)ethash', 'High'),
-        'kawpow': (rb'(?i)kawpow', 'High'),
-        'equihash': (rb'(?i)equihash', 'High'),
-        'scrypt': (rb'(?i)(scrypt|litecoin)', 'High'),
-        
-        # Cryptocurrency Wallets in Memory
-        'monero_wallet': (rb'(?i)(4[0-9AB][1-9A-HJ-NP-Za-km-z]{93})', 'Critical'),
-        'bitcoin_wallet': (rb'(?i)([13][a-km-zA-HJ-NP-Z1-9]{25,34})', 'High'),
-        'ethereum_wallet': (rb'(?i)(0x[a-fA-F0-9]{40})', 'High'),
-        
-        # Mining Configuration
-        'mining_config': (rb'(?i)(mining.*config|config.*pool|worker.*name)', 'Medium'),
-        'hashrate': (rb'(?i)(hashrate|h/s|kh/s|mh/s)', 'Medium'),
-        'difficulty': (rb'(?i)(difficulty.*target|share.*difficulty)', 'Medium'),
-        
-        # Malicious Miners
-        'coinhive': (rb'(?i)(coinhive|coin-hive)', 'Critical'),
-        'cryptoloot': (rb'(?i)cryptoloot', 'Critical'),
-        'jsecoin': (rb'(?i)jsecoin', 'Critical'),
-        'webminerpool': (rb'(?i)webminerpool', 'Critical'),
-        'mineralt': (rb'(?i)mineralt', 'Critical'),
-        'authedmine': (rb'(?i)authedmine', 'Critical'),
-        'coinimp': (rb'(?i)coinimp', 'Critical'),
-        'crypto_loot': (rb'(?i)crypto-loot', 'Critical'),
-        'deepminer': (rb'(?i)deepminer', 'Critical'),
-        'gridcash': (rb'(?i)gridcash', 'Critical'),
-        
-        # Hidden Miners
-        'xmr_stak': (rb'(?i)(xmr-stak|xmrstak)', 'Critical'),
-        'srbminer': (rb'(?i)srbminer', 'Critical'),
-        'wildrig': (rb'(?i)wildrig', 'Critical'),
-        'kawpowminer': (rb'(?i)kawpowminer', 'Critical'),
-        
-        # Browser-based Mining
-        'browser_mining': (rb'(?i)(cryptonight\.wasm|mining\.js|miner\.js)', 'High'),
-        'web_assembly_mining': (rb'(?i)(wasm.*crypto|crypto.*wasm)', 'High'),
-        
-        # Mining Commands
-        'mining_command': (rb'(?i)(--algo|--pool|--user|--pass.*x|--donate-level)', 'High'),
-        'gpu_mining': (rb'(?i)(--cuda|--opencl|--gpu)', 'Medium'),
-        'cpu_mining': (rb'(?i)(--threads|--cpu-priority)', 'Medium'),
-        
-        # Stealth Mining Indicators
-        'process_hiding': (rb'(?i)(hide.*window|invisible.*mode|stealth.*mode)', 'High'),
-        'mining_rootkit': (rb'(?i)(rootkit.*miner|miner.*rootkit)', 'Critical'),
-        'persistence_mining': (rb'(?i)(scheduled.*miner|startup.*miner)', 'High'),
-    }
-    
-    # Mining Process Names
-    MINER_PROCESS_NAMES = {
-        'xmrig.exe': 'XMRig Monero Miner',
-        'xmrig-cuda.exe': 'XMRig CUDA Miner',
-        'xmrig-nvidia.exe': 'XMRig NVIDIA Miner',
-        'xmrig-amd.exe': 'XMRig AMD Miner',
-        'claymore.exe': 'Claymore Miner',
-        'ethminer.exe': 'Ethereum Miner',
-        'phoenixminer.exe': 'Phoenix Miner',
-        'cgminer.exe': 'CGMiner',
-        'bfgminer.exe': 'BFGMiner',
-        'nheqminer.exe': 'NiceHash Miner',
-        'ccminer.exe': 'CCMiner',
-        'minerd.exe': 'CPU Miner',
-        'cpuminer.exe': 'CPU Miner',
-        'minergate.exe': 'MinerGate',
-        'nicehash.exe': 'NiceHash',
-        'gminer.exe': 'GMiner',
-        'lolminer.exe': 'LolMiner',
-        'nbminer.exe': 'NBMiner',
-        't-rex.exe': 'T-Rex Miner',
-        'teamredminer.exe': 'Team Red Miner',
-        'srbminer.exe': 'SRBMiner',
-        'wildrig.exe': 'WildRig Miner',
-        'bminer.exe': 'BMiner',
-        'ewbf.exe': 'EWBF Miner',
-        'excavator.exe': 'NiceHash Excavator',
-        'xmr-stak.exe': 'XMR-Stak',
-        'xmr-stack-cpu.exe': 'XMR-Stak CPU',
-        'xmr-stack-amd.exe': 'XMR-Stak AMD',
-        'xmr-stack-nvidia.exe': 'XMR-Stak NVIDIA',
-    }
-    
-    # Remote Access Trojans (RATs) signatures
-    RAT_SIGNATURES = {
-        # njRAT - Multiple signatures for better detection
-        'njrat': (rb'(?i)(njrat|bladabindi)', 'Critical'),
-        'njrat_mutex': (rb'(?i)(njrat.*mutex|bladabindi.*mutex)', 'Critical'),
-        'njrat_registry': (rb'(?i)(njrat.*registry|software\\microsoft\\windows\\currentversion\\run.*njrat)', 'Critical'),
-        'njrat_c2': (rb'(?i)(njrat.*server|njrat.*c2|njrat.*panel)', 'Critical'),
-        'njrat_plugin': (rb'(?i)(plugin.*njrat|njrat.*plugin)', 'Critical'),
-        'njrat_strings': (rb'(?i)(njworm|njspy|njlogger)', 'Critical'),
-        'njrat_commands': (rb'(?i)(ll|rn|inv|ret|CAP|un|up|RG|kl)', 'High'),
-        'njrat_base64': (rb'(?i)(am5yYXQ|bmpy|Ymxh)', 'High'),  # njrat base64 encoded
-        
-        # DarkComet - Enhanced detection
-        'darkcomet': (rb'(?i)(darkcomet|dc_mutexrat)', 'Critical'),
-        'darkcomet_mutex': (rb'(?i)(dc_mutex|dcmutex|darkcomet.*mutex)', 'Critical'),
-        'darkcomet_registry': (rb'(?i)(darkcomet.*registry|dcregistry)', 'Critical'),
-        'darkcomet_rat': (rb'(?i)(dcrat|darkcomet.*rat)', 'Critical'),
-        'darkcomet_version': (rb'(?i)(darkcomet.*5\.[0-9]|dc.*5\.[0-9])', 'Critical'),
-        'darkcomet_keylogger': (rb'(?i)(dc.*keylog|darkcomet.*keylog)', 'Critical'),
-        'darkcomet_stub': (rb'(?i)(stub.*darkcomet|dc.*stub)', 'Critical'),
-        'darkcomet_strings': (rb'(?i)(DCRATKILL|DCRATID|GENCODE)', 'Critical'),
-        'darkcomet_config': (rb'(?i)(darkcomet.*config|dcconfig)', 'High'),
-        
-        # Quasar RAT - Comprehensive signatures
-        'quasar': (rb'(?i)(quasar.*rat|xrat)', 'Critical'),
-        'quasar_namespace': (rb'(?i)(xrat\..*|quasar\.)', 'Critical'),
-        'quasar_mutex': (rb'(?i)(quasar.*mutex|xrat.*mutex)', 'Critical'),
-        'quasar_client': (rb'(?i)(quasar.*client|quasarclient)', 'Critical'),
-        'quasar_server': (rb'(?i)(quasar.*server|quasarserver)', 'Critical'),
-        'quasar_assembly': (rb'(?i)(quasar\..*assembly)', 'Critical'),
-        'quasar_keylogger': (rb'(?i)(quasar.*keylogger)', 'Critical'),
-        'quasar_commands': (rb'(?i)(getpasswords|getserverpassword|remotedesktop)', 'High'),
-        'quasar_config': (rb'(?i)(quasar.*settings|quasar.*config)', 'High'),
-        'quasar_communication': (rb'(?i)(quasar.*networking|xrat.*net)', 'High'),
-        
-        # Popular RATs
-        'nanocore': (rb'(?i)(nanocore|nanobot)', 'Critical'),
-        'remcos': (rb'(?i)(remcos|remcosrat)', 'Critical'),
-        'asyncrat': (rb'(?i)(asyncrat|async.*rat)', 'Critical'),
-        'netwire': (rb'(?i)(netwire|netwiredrc)', 'Critical'),
-        'poisonivy': (rb'(?i)(poison.*ivy|pivy)', 'Critical'),
-        'blackshades': (rb'(?i)(blackshades|bsnet)', 'Critical'),
-        'cybergate': (rb'(?i)cybergate', 'Critical'),
-        
-        # Advanced RATs
-        'cobaltstrike': (rb'(?i)(beacon.*dll|cobaltstrike|malleable)', 'Critical'),
-        'meterpreter': (rb'(?i)(meterpreter|msf.*payload)', 'Critical'),
-        'empire': (rb'(?i)(powershell.*empire|invoke.*empire)', 'Critical'),
-        'pupy': (rb'(?i)pupy.*rat', 'Critical'),
-        'covenant': (rb'(?i)(covenant|grunt)', 'Critical'),
-        
-        # Commercial/Government RATs
-        'gh0st': (rb'(?i)(gh0st|gh0strat)', 'Critical'),
-        'plugx': (rb'(?i)(plugx|korplug)', 'Critical'),
-        'sakula': (rb'(?i)sakula', 'Critical'),
-        'winnti': (rb'(?i)winnti', 'Critical'),
-        'shadowpad': (rb'(?i)shadowpad', 'Critical'),
-        'covenant_rat': (rb'(?i)covenant', 'Critical'),
-        
-        # Mobile/Cross-platform RATs
-        'androrat': (rb'(?i)androrat', 'Critical'),
-        'omnirat': (rb'(?i)omnirat', 'Critical'),
-        'spynote': (rb'(?i)spynote', 'Critical'),
-        
-        # Banking/Stealer RATs
-        'zeus': (rb'(?i)(zeus|zbot)', 'Critical'),
-        'carberp': (rb'(?i)carberp', 'Critical'),
-        'tinba': (rb'(?i)tinba', 'Critical'),
-        'ursnif': (rb'(?i)(ursnif|gozi)', 'Critical'),
-        'formbook': (rb'(?i)formbook', 'Critical'),
-        'agent_tesla': (rb'(?i)(agenttesla|agent.*tesla)', 'Critical'),
-        'lokibot': (rb'(?i)(lokibot|loki.*pwd)', 'Critical'),
-        'azorult': (rb'(?i)azorult', 'Critical'),
-        'raccoon': (rb'(?i)(raccoon.*stealer)', 'Critical'),
-        
-        # APT-related RATs
-        'htran': (rb'(?i)htran', 'Critical'),
-        'reaver': (rb'(?i)reaver', 'Critical'),
-        'cobalt_kitty': (rb'(?i)cobalt.*kitty', 'Critical'),
-        'darkhotel': (rb'(?i)darkhotel', 'Critical'),
-        'apt1_rat': (rb'(?i)(seasalt|rockboot)', 'Critical'),
-        
-        # Open Source RATs
-        'orcus': (rb'(?i)orcus.*rat', 'Critical'),
-        'revenge_rat': (rb'(?i)revenge.*rat', 'Critical'),
-        'havex': (rb'(?i)havex', 'Critical'),
-        'adwind': (rb'(?i)(adwind|jrat|sockrat)', 'Critical'),
-        'unrecom': (rb'(?i)unrecom', 'Critical'),
-        'njw0rm': (rb'(?i)njw0rm', 'Critical'),
-        'lime_rat': (rb'(?i)lime.*rat', 'Critical'),
-        'xpertrat': (rb'(?i)xpert.*rat', 'Critical'),
-        
-        # Modern RATs (2020+)
-        'sliver': (rb'(?i)sliver', 'Critical'),
-        'brute_ratel': (rb'(?i)(brute.*ratel|badger)', 'Critical'),
-        'mythic': (rb'(?i)(mythic|apollo.*agent)', 'Critical'),
-        'havoc': (rb'(?i)havoc.*c2', 'Critical'),
-        'villain': (rb'(?i)villain.*c2', 'Critical'),
-        
-        # Backdoors
-        'backdoor_factory': (rb'(?i)(backdoor.*factory|bdf)', 'Critical'),
-        'bifrost': (rb'(?i)bifrost', 'Critical'),
-        'turla': (rb'(?i)(turla|uroburos)', 'Critical'),
-        'derusbi': (rb'(?i)derusbi', 'Critical'),
-        
-        # Web Shells
-        'china_chopper': (rb'(?i)(china.*chopper|caidao)', 'Critical'),
-        'aspxspy': (rb'(?i)aspxspy', 'Critical'),
-        'webshell': (rb'(?i)(eval.*request|execute.*request)', 'High'),
-        
-        # Commodity RATs
-        'warzone': (rb'(?i)(warzone.*rat|ave.*maria)', 'Critical'),
-        'dcrat': (rb'(?i)dcrat', 'Critical'),
-        'parallax': (rb'(?i)parallax.*rat', 'Critical'),
-        'xworm': (rb'(?i)xworm', 'Critical'),
-        'plasma': (rb'(?i)plasma.*rat', 'Critical'),
-    }
-    
-    # Network indicators
-    NETWORK_PATTERNS = {
-        'c2_beacon': (rb'(?i)(beacon|heartbeat|checkin)', 'High'),
-        'reverse_shell': (rb'(?i)(reverse.*shell|bind.*shell)', 'Critical'),
-        'tunnel': (rb'(?i)(chisel|ngrok|serveo)', 'High'),
-        'exfiltration': (rb'(?i)(exfil|upload.*data)', 'Critical'),
-    }
-    
-    # Living off the land binaries with enhanced detection
-    LOLBINS = {
-        'powershell.exe': ('PowerShell', 'T1059.001'),
-        'pwsh.exe': ('PowerShell Core', 'T1059.001'),
-        'powershell_ise.exe': ('PowerShell ISE', 'T1059.001'),
-        'cmd.exe': ('Command Prompt', 'T1059.003'),
-        'wmic.exe': ('WMI Command', 'T1047'),
-        'mshta.exe': ('HTML Application', 'T1218.005'),
-        'regsvr32.exe': ('Register Server', 'T1218.010'),
-        'rundll32.exe': ('Run DLL', 'T1218.011'),
-        'cscript.exe': ('VBScript Engine', 'T1059.005'),
-        'wscript.exe': ('Windows Script Host', 'T1059.005'),
-        'certutil.exe': ('Certificate Utility', 'T1140'),
-        'bitsadmin.exe': ('BITS Admin', 'T1197'),
-        'msiexec.exe': ('Windows Installer', 'T1218.007'),
-        'regasm.exe': ('.NET Assembly Registration', 'T1218.009'),
-        'regsvcs.exe': ('.NET Services', 'T1218.009'),
-        'installutil.exe': ('.NET Installer', 'T1218.004'),
-        'msbuild.exe': ('MSBuild', 'T1127.001'),
-        'cmstp.exe': ('Connection Manager', 'T1218.003'),
-        'odbcconf.exe': ('ODBC Config', 'T1218.008'),
-        'schtasks.exe': ('Task Scheduler', 'T1053.005'),
-        'at.exe': ('AT Command', 'T1053.002'),
-        'sc.exe': ('Service Control', 'T1543.003'),
-        'net.exe': ('Net Command', 'T1087'),
-        'net1.exe': ('Net Command', 'T1087'),
-        'whoami.exe': ('User Identity', 'T1033'),
-        'systeminfo.exe': ('System Info', 'T1082'),
-        'tasklist.exe': ('Task List', 'T1057'),
-        'netstat.exe': ('Network Stats', 'T1049'),
-        'ipconfig.exe': ('IP Config', 'T1016'),
-        'arp.exe': ('ARP', 'T1018'),
-        'route.exe': ('Route', 'T1016'),
-        'netsh.exe': ('Network Shell', 'T1562.004'),
-        'forfiles.exe': ('ForFiles', 'T1222'),
-        'pcalua.exe': ('Program Compatibility', 'T1218'),
-        'mavinject.exe': ('Mavinject', 'T1218'),
-        'control.exe': ('Control Panel', 'T1218'),
-        'msdt.exe': ('Microsoft Diagnostics', 'T1218'),
-    }
-    
-    # Suspicious DLLs
-    SUSPICIOUS_DLLS = [
-        'sbiedll.dll', 'dbghelp.dll', 'api_log.dll', 'vmGuestLib.dll',
-        'vboxmrxnp.dll', 'VBoxHook.dll', 'prltools.dll', 'inject.dll',
-        'reflective.dll', 'magic.dll', 'hook.dll'
-    ]
-    
-    # Code cave indicators (RWX memory)
-    CODE_CAVE_SIZE_THRESHOLD = 4096  # Increased from 1024 to reduce false positives
-    
-    # Minimum match confidence
-    MIN_PATTERN_LENGTH = 20  # Minimum bytes to match for detection
-    
-    # Whitelist for known legitimate processes
-    LEGITIMATE_PROCESSES = [
-        'system', 'smss.exe', 'csrss.exe', 'wininit.exe', 'services.exe',
-        'lsass.exe', 'svchost.exe', 'winlogon.exe', 'dwm.exe', 'explorer.exe',
-        'taskhost.exe', 'taskhostw.exe', 'spoolsv.exe', 'sihost.exe'
-    ]
-    
-    @staticmethod
-    def _is_false_positive(proc_name: str, detection_type: str, indicator: str) -> bool:
-        """Check if detection is likely a false positive."""
-        
-        # Skip detections in core system processes for certain types
-        if proc_name in FilelessMalwareHunter.LEGITIMATE_PROCESSES:
-            # Allow only critical detections in system processes
-            if detection_type in ['LOLBin Abuse', '.NET Injection', 'Orphan Process']:
-                return True
-        
-        # Filter out common false positives
-        false_positive_patterns = [
-            # Common legitimate .NET usage
-            (detection_type == '.NET Injection' and proc_name in ['explorer.exe', 'taskhost.exe', 'taskhostw.exe']),
-            
-            # Legitimate svchost instances
-            (detection_type == 'Orphan Process' and proc_name == 'svchost.exe'),
-            
-            # System processes with low thread counts
-            (detection_type == 'Thread Injection' and proc_name in FilelessMalwareHunter.LEGITIMATE_PROCESSES),
-        ]
-        
-        return any(false_positive_patterns)
-    
-    @staticmethod
-    def _validate_detection(data: bytes, pattern: bytes, min_length: int = 20) -> bool:
-        """Validate that pattern match is substantial enough."""
-        match = re.search(pattern, data)
-        if match:
-            matched_text = match.group(0)
-            # Ensure matched text is substantial
-            return len(matched_text) >= min_length
-        return False
-    
-    @staticmethod
-    def _get_severity_value(severity: str) -> int:
-        """Convert severity string to numeric value."""
-        severity_map = {'Low': 1, 'Medium': 2, 'High': 3, 'Critical': 4}
-        return severity_map.get(severity, 1)
-    
-    @staticmethod
-    def scan_process_memory(pid: int = None):
-        """Scan process memory for suspicious patterns with enhanced detection."""
-        detections = []
-        
-        try:
-            processes = []
-            if pid:
-                try:
-                    proc = psutil.Process(pid)
-                    processes.append(proc)
-                except psutil.NoSuchProcess:
-                    return detections
-            else:
-                processes = list(psutil.process_iter())
-            
-            for proc in processes:
-                try:
-                    proc_name = proc.name()
-                    proc_pid = proc.pid
-                    
-                    # Skip system processes for certain checks
-                    if proc_name in FilelessMalwareHunter.LEGITIMATE_PROCESSES:
-                        continue
-                    
-                    # Check for known miner processes
-                    if proc_name.lower() in FilelessMalwareHunter.MINER_PROCESS_NAMES:
-                        detections.append({
-                            'type': 'Cryptocurrency Miner',
-                            'process': proc_name,
-                            'pid': proc_pid,
-                            'indicator': f'Known miner process: {FilelessMalwareHunter.MINER_PROCESS_NAMES[proc_name.lower()]}',
-                            'severity': 'Critical',
-                            'address': 'N/A',
-                            'protection': 'N/A',
-                            'technique': 'T1496 - Resource Hijacking'
-                        })
-                    
-                    # Check command line for suspicious patterns
-                    try:
-                        cmdline = ' '.join(proc.cmdline()) if proc.cmdline() else ''
-                        
-                        # Check for PowerShell patterns in command line
-                        for pattern_name, (pattern, severity) in FilelessMalwareHunter.POWERSHELL_PATTERNS.items():
-                            if re.search(pattern, cmdline.encode('utf-8', errors='ignore')):
-                                detections.append({
-                                    'type': 'PowerShell Script',
-                                    'process': proc_name,
-                                    'pid': proc_pid,
-                                    'indicator': pattern_name.replace('_', ' ').title(),
-                                    'severity': severity,
-                                    'address': cmdline[:100] + '...' if len(cmdline) > 100 else cmdline,
-                                    'protection': 'N/A',
-                                    'technique': 'T1059.001 - PowerShell'
-                                })
-                                break
-                        
-                        # Check for LOLBin abuse in command line
-                        if proc_name in FilelessMalwareHunter.LOLBINS:
-                            lolbin_info = FilelessMalwareHunter.LOLBINS[proc_name]
-                            suspicious_patterns = {
-                                'http://': ('Remote HTTP Download', 'T1105 - Ingress Tool Transfer', 'Critical'),
-                                'https://': ('Remote HTTPS Download', 'T1105 - Ingress Tool Transfer', 'High'),
-                                'ftp://': ('FTP Transfer', 'T1105 - Ingress Tool Transfer', 'High'),
-                                '-enc': ('Encoded Command', 'T1027 - Obfuscated Files', 'High'),
-                                '-e ': ('Encoded Command', 'T1027 - Obfuscated Files', 'High'),
-                                'frombase64': ('Base64 Decode', 'T1140 - Deobfuscate/Decode', 'High'),
-                                'downloadstring': ('Web Download', 'T1105 - Ingress Tool Transfer', 'Critical'),
-                                'downloadfile': ('File Download', 'T1105 - Ingress Tool Transfer', 'Critical'),
-                                'invoke-expression': ('Dynamic Execution', 'T1059.001 - PowerShell', 'High'),
-                                'iex': ('Dynamic Execution', 'T1059.001 - PowerShell', 'High'),
-                                'bypass': ('Execution Policy Bypass', 'T1562.001 - Disable/Modify Tools', 'High'),
-                                '-nop': ('No Profile', 'T1562.001 - Disable/Modify Tools', 'Medium'),
-                                '-w hidden': ('Hidden Window', 'T1564.003 - Hidden Window', 'High'),
-                                'windowstyle hidden': ('Hidden Window', 'T1564.003 - Hidden Window', 'High'),
-                                'amsi': ('AMSI Bypass', 'T1562.001 - Disable/Modify Tools', 'Critical'),
-                            }
-                            
-                            for pattern, (desc, technique, severity) in suspicious_patterns.items():
-                                if pattern.lower() in cmdline.lower():
-                                    detections.append({
-                                        'type': 'LOLBin Abuse',
-                                        'process': proc_name,
-                                        'pid': proc_pid,
-                                        'indicator': f'{lolbin_info[0]}: {desc}',
-                                        'severity': severity,
-                                        'address': cmdline[:150] + '...' if len(cmdline) > 150 else cmdline,
-                                        'protection': 'N/A',
-                                        'technique': technique
-                                    })
-                                    break
-                    
-                    except (psutil.AccessDenied, psutil.NoSuchProcess):
-                        pass
-                    
-                    # Check for suspicious parent-child relationships
-                    try:
-                        parent = proc.parent()
-                        if parent:
-                            parent_name = parent.name()
-                            
-                            suspicious_relations = {
-                                'winword.exe': {
-                                    'children': ['powershell.exe', 'cmd.exe', 'wmic.exe', 'mshta.exe', 'certutil.exe', 'regsvr32.exe'],
-                                    'technique': 'T1566.001 - Spearphishing Attachment',
-                                    'severity': 'Critical'
-                                },
-                                'excel.exe': {
-                                    'children': ['powershell.exe', 'cmd.exe', 'wmic.exe', 'certutil.exe', 'regsvr32.exe'],
-                                    'technique': 'T1566.001 - Spearphishing Attachment',
-                                    'severity': 'Critical'
-                                },
-                                'outlook.exe': {
-                                    'children': ['powershell.exe', 'cmd.exe', 'wscript.exe', 'cscript.exe'],
-                                    'technique': 'T1566.002 - Spearphishing Link',
-                                    'severity': 'High'
-                                },
-                                'acrord32.exe': {
-                                    'children': ['powershell.exe', 'cmd.exe', 'wscript.exe'],
-                                    'technique': 'T1204.002 - Malicious File',
-                                    'severity': 'High'
-                                },
-                            }
-                            
-                            for parent_pattern, config in suspicious_relations.items():
-                                if parent_pattern in parent_name and proc_name in config['children']:
-                                    detections.append({
-                                        'type': 'Suspicious Parent-Child',
-                                        'process': proc_name,
-                                        'pid': proc_pid,
-                                        'indicator': f'{parent_name} → {proc_name}',
-                                        'severity': config['severity'],
-                                        'address': f'PPID: {parent.pid}',
-                                        'protection': 'N/A',
-                                        'technique': config['technique']
-                                    })
-                    
-                    except (psutil.AccessDenied, psutil.NoSuchProcess):
-                        pass
-                    
-                    # Check for orphan processes
-                    try:
-                        parent = proc.parent()
-                        if not parent and proc_name not in FilelessMalwareHunter.LEGITIMATE_PROCESSES:
-                            detections.append({
-                                'type': 'Orphan Process',
-                                'process': proc_name,
-                                'pid': proc_pid,
-                                'indicator': 'No parent process found',
-                                'severity': 'Medium',
-                                'address': 'N/A',
-                                'protection': 'N/A',
-                                'technique': 'T1055 - Process Injection'
-                            })
-                    
-                    except (psutil.AccessDenied, psutil.NoSuchProcess):
-                        pass
-                    
-                    # Check memory usage for suspicious patterns (high memory usage for certain processes)
-                    try:
-                        memory_info = proc.memory_info()
-                        if memory_info.rss > 500 * 1024 * 1024:  # 500 MB
-                            if proc_name in ['powershell.exe', 'cmd.exe', 'wscript.exe', 'cscript.exe']:
-                                detections.append({
-                                    'type': 'Suspicious Memory Usage',
-                                    'process': proc_name,
-                                    'pid': proc_pid,
-                                    'indicator': f'High memory usage: {memory_info.rss / (1024*1024):.2f} MB',
-                                    'severity': 'Medium',
-                                    'address': 'N/A',
-                                    'protection': 'N/A',
-                                    'technique': 'T1055 - Process Injection'
-                                })
-                    
-                    except (psutil.AccessDenied, psutil.NoSuchProcess):
-                        pass
-                        
-                except (psutil.NoSuchProcess, psutil.AccessDenied):
-                    continue
-                    
-        except Exception as e:
-            vollog.debug(f"Error in fileless malware scanning: {e}")
-        
-        return detections
-    
-    @staticmethod
-    def scan_system():
-        """Perform comprehensive fileless malware scan."""
-        detections = FilelessMalwareHunter.scan_process_memory()
-        
-        # Add additional system-wide checks
-        try:
-            # Check for suspicious scheduled tasks
-            tasks = LiveSystemContext.get_scheduled_tasks_registry()
-            for task in tasks:
-                task_name = task['Name'].lower()
-                task_actions = task['Actions'].lower()
-                
-                # Check for suspicious task names and actions
-                suspicious_tasks = {
-                    'update': 'Disguised as system update',
-                    'microsoft': 'Impersonating Microsoft',
-                    'java': 'Impersonating Java',
-                    'adobe': 'Impersonating Adobe',
-                    'google': 'Impersonating Google',
-                }
-                
-                for pattern, description in suspicious_tasks.items():
-                    if pattern in task_name and any(cmd in task_actions for cmd in ['powershell', 'cmd', 'wscript']):
-                        detections.append({
-                            'type': 'Suspicious Scheduled Task',
-                            'process': 'Task Scheduler',
-                            'pid': 0,
-                            'indicator': f'{description}: {task["Name"]}',
-                            'severity': 'High',
-                            'address': task['Actions'],
-                            'protection': 'N/A',
-                            'technique': 'T1053.005 - Scheduled Task'
-                        })
-                        break
-            
-            # Check for persistence mechanisms
-            persistence = LiveSystemContext.get_persistence_mechanisms()
-            for item in persistence:
-                if any(suspicious in item['Target'].lower() for suspicious in ['powershell', 'cmd', 'wscript', 'cscript', 'mshta']):
-                    detections.append({
-                        'type': 'Persistence Mechanism',
-                        'process': 'System',
-                        'pid': 0,
-                        'indicator': f'{item["Type"]}: {item["Name"]}',
-                        'severity': 'High',
-                        'address': item['Target'],
-                        'protection': 'N/A',
-                        'technique': 'T1547 - Boot/Logon Autostart'
-                    })
-        
-        except Exception as e:
-            vollog.debug(f"Error in system-wide fileless scan: {e}")
-        
-        return detections
-
-
-class LiveShellCommand(cmd.Cmd):
-    """Interactive shell for live system analysis."""
-    
-    intro = ""
-    prompt = '\033[92mvol3-live>\033[0m '
-    
-    def __init__(self):
-        super().__init__()
-        self.context = LiveSystemContext()
-        self.last_result = None
-        self._intro_shown = True
-    
-    def preloop(self):
-        """Override preloop to prevent showing intro."""
-        pass
-
-    # NEW PLUGIN METHODS
-    def do_timeliner(self, arg):
-        """Create timeline forensics analysis of system artifacts
-        Usage: timeliner [--limit <n>] [--source <process|file|registry|network|prefetch|eventlog>]"""
-        
-        args = self._parse_args(arg)
-        limit = int(args.get('--limit', 100))
-        source_filter = args.get('--source', '').lower()
-        
-        print("\n[*] Collecting timeline forensic artifacts...")
-        print("[*] Sources: Process creation, File system, Registry, Network, Prefetch, Event logs")
-        
-        timeline_data = self.context.get_timeline_artifacts()
-        
-        if source_filter:
-            timeline_data = [item for item in timeline_data if source_filter in item['Source'].lower()]
-        
-        if not timeline_data:
-            print(f"\nNo timeline data found{' matching source filter' if source_filter else ''}\n")
-            return
-        
-        print(f"\nTimeline Forensics Analysis (Recent {limit} events):")
-        print("=" * 150)
-        print("{:<20} {:<15} {:<25} {:<80} {:<8}".format(
-            "Timestamp", "Source", "Action", "Details", "PID"))
-        print("=" * 150)
-        
-        for event in timeline_data[:limit]:
-            print("{:<20} {:<15} {:<25} {:<80} {:<8}".format(
-                event['Timestamp'][:19],
-                event['Source'][:14],
-                event['Action'][:24],
-                event['Details'][:79],
-                event['PID']
-            ))
-        
-        if len(timeline_data) > limit:
-            print(f"\n... and {len(timeline_data) - limit} more events")
-        print(f"\nTotal timeline events: {len(timeline_data)}")
-        print("\nTIP: Use --source to filter by: process, file, registry, network, prefetch, eventlog\n")
-    
-    def do_disklist(self, arg):
-        """Display hard drives and storage devices information
-        Usage: disklist [--type <physical|logical|partition|usb>]"""
-        
-        args = self._parse_args(arg)
-        type_filter = args.get('--type', '').lower()
-        
-        print("\n[*] Scanning for storage devices...")
-        disk_info = self.context.get_disk_list()
-        
-        if type_filter:
-            disk_info = [disk for disk in disk_info if type_filter in disk['Type'].lower()]
-        
-        if not disk_info:
-            print(f"\nNo disks found{' matching type filter' if type_filter else ''}\n")
-            return
-        
-        print("\nStorage Devices Information:")
-        print("=" * 180)
-        print("{:<15} {:<20} {:<10} {:<15} {:<15} {:<15} {:<12} {:<20} {:<50}".format(
-            "Device", "Mount Point", "Type", "File System", "Total Size", "Used", "Free", "Percent Used", "Additional Info"))
-        print("=" * 180)
-        
-        for disk in disk_info:
-            additional_info = ""
-            if 'Model' in disk and disk['Model'] != 'N/A':
-                additional_info = f"Model: {disk['Model']}"
-            if 'Interface' in disk and disk['Interface'] != 'N/A':
-                additional_info += f", Interface: {disk['Interface']}"
-            if 'VolumeName' in disk and disk['VolumeName'] != 'N/A':
-                additional_info = f"Volume: {disk['VolumeName']}"
-            
-            print("{:<15} {:<20} {:<10} {:<15} {:<15} {:<15} {:<12} {:<20} {:<50}".format(
-                disk['Device'][:14],
-                disk['MountPoint'][:19],
-                disk['Type'][:9],
-                disk['FileSystem'][:14],
-                disk['TotalSize'][:14],
-                disk['Used'][:14],
-                disk['Free'][:14],
-                disk['PercentUsed'][:11],
-                additional_info[:49]
-            ))
-        
-        print(f"\nTotal storage devices: {len(disk_info)}")
-        print("\nTIP: Use --type to filter by: physical, logical, partition, usb\n")
-    
-    def do_sandbox(self, arg):
-        """Check if system is running in VM/sandbox
-        Usage: sandbox"""
-        
-        print("\n[*] Performing sandbox/VM detection...")
-        indicators = self.context.detect_sandbox()
-        
-        if not indicators:
-            print("\n[+] No strong VM/sandbox indicators detected")
-            print("[-] System appears to be physical hardware")
-        else:
-            print(f"\n[!] {len(indicators)} VM/sandbox indicators detected:")
-            print("=" * 80)
-            for indicator in indicators:
-                print(f"  • {indicator}")
-        
-        print()
-    
-    def do_prefetch(self, arg):
-        """Display Prefetch files information
-        Usage: prefetch [--limit <n>]"""
-        
-        args = self._parse_args(arg)
-        limit = int(args.get('--limit', 50))
-        
-        print("\n[*] Retrieving Prefetch files...")
-        prefetch_files = self.context.get_prefetch_files()
-        
-        if not prefetch_files:
-            print("\nNo Prefetch files found or access denied\n")
-            return
-        
-        print(f"\nPrefetch Files (Recent {limit}):")
-        print("=" * 140)
-        print("{:<40} {:<12} {:<20} {:<20} {:<20} {:<20}".format(
-            "Filename", "Size", "Executable", "Created", "Modified", "Accessed"))
-        print("=" * 140)
-        
-        for pf in prefetch_files[:limit]:
-            size_str = f"{pf['Size'] / 1024:.1f} KB"
-            exe_name = pf.get('Executable', 'Unknown')
-            
-            print("{:<40} {:<12} {:<20} {:<20} {:<20} {:<20}".format(
-                pf['Filename'][:39],
-                size_str,
-                exe_name[:19],
-                pf['Created'][:19],
-                pf['Modified'][:19],
-                pf['Accessed'][:19]
-            ))
-        
-        if len(prefetch_files) > limit:
-            print(f"\n... and {len(prefetch_files) - limit} more prefetch files")
-        print(f"\nTotal prefetch files: {len(prefetch_files)}\n")
-    
-    def do_shellbags(self, arg):
-        """Display Shellbags information (folder navigation history)
-        Usage: shellbags [--limit <n>]"""
-        
-        args = self._parse_args(arg)
-        limit = int(args.get('--limit', 100))
-        
-        print("\n[*] Retrieving Shellbags from registry...")
-        shellbags = self.context.get_shellbags()
-        
-        if not shellbags:
-            print("\nNo Shellbags found or access denied\n")
-            return
-        
-        print(f"\nShellbags (Folder Navigation History):")
-        print("=" * 120)
-        print("{:<60} {:<15} {:<20} {:<20}".format(
-            "Registry Path", "Data Size", "Data Type", "Target Path"))
-        print("=" * 120)
-        
-        for sb in shellbags[:limit]:
-            target = sb.get('Target', 'N/A')[:39] if 'Target' in sb else 'N/A'
-            print("{:<60} {:<15} {:<20} {:<20}".format(
-                sb['Path'][:59],
-                f"{sb.get('DataSize', 0)} bytes",
-                sb.get('DataType', 'Unknown')[:19],
-                target
-            ))
-        
-        if len(shellbags) > limit:
-            print(f"\n... and {len(shellbags) - limit} more shellbag entries")
-        print(f"\nTotal shellbag entries: {len(shellbags)}\n")
-    
-    def do_amcache(self, arg):
-        """Display Amcache.hve information (application compatibility cache)
-        Usage: amcache"""
-        
-        print("\n[*] Retrieving Amcache data...")
-        amcache_entries = self.context.get_amcache()
-        
-        if not amcache_entries:
-            print("\nNo Amcache data found or access denied\n")
-            return
-        
-        print("\nAmcache (Application Compatibility Cache):")
-        print("=" * 120)
-        print("{:<20} {:<60} {:<12} {:<20} {:<20}".format(
-            "Type", "Path", "Size", "Modified", "Status"))
-        print("=" * 120)
-        
-        for entry in amcache_entries:
-            size_str = f"{entry['Size'] / 1024:.1f} KB" if entry['Size'] > 0 else "N/A"
-            print("{:<20} {:<60} {:<12} {:<20} {:<20}".format(
-                entry['Type'][:19],
-                entry['Path'][:59],
-                size_str,
-                entry.get('Modified', 'N/A')[:19],
-                entry.get('Status', 'N/A')[:19]
-            ))
-        
-        print(f"\nTotal Amcache entries: {len(amcache_entries)}\n")
-        print("Note: Full Amcache parsing requires specialized registry hive parsing tools")
-    
-    def do_shimcache(self, arg):
-        """Display Shimcache information (application compatibility cache)
-        Usage: shimcache [--limit <n>]"""
-        
-        args = self._parse_args(arg)
-        limit = int(args.get('--limit', 100))
-        
-        print("\n[*] Retrieving Shimcache data...")
-        shimcache_entries = self.context.get_shimcache()
-        
-        if not shimcache_entries:
-            print("\nNo Shimcache data found or access denied\n")
-            return
-        
-        print(f"\nShimcache (Application Compatibility Cache):")
-        print("=" * 120)
-        print("{:<60} {:<20} {:<15} {:<20}".format(
-            "Registry Path", "Value Name", "Data Size", "File Path"))
-        print("=" * 120)
-        
-        for sc in shimcache_entries[:limit]:
-            value_name = sc.get('ValueName', 'N/A')[:19]
-            filepath = sc.get('Filepath', 'N/A')[:39]
-            
-            print("{:<60} {:<20} {:<15} {:<20}".format(
-                sc['RegistryPath'][:59],
-                value_name,
-                f"{sc.get('DataSize', 0)} bytes",
-                filepath
-            ))
-        
-        if len(shimcache_entries) > limit:
-            print(f"\n... and {len(shimcache_entries) - limit} more shimcache entries")
-        print(f"\nTotal shimcache entries: {len(shimcache_entries)}\n")
-    
-    def do_jumplists(self, arg):
-        """Display Jump List information (recent documents and tasks)
-        Usage: jumplists [--limit <n>]"""
-        
-        args = self._parse_args(arg)
-        limit = int(args.get('--limit', 50))
-        
-        print("\n[*] Retrieving Jump Lists...")
-        jumplists = self.context.get_jumplists()
-        
-        if not jumplists:
-            print("\nNo Jump Lists found or access denied\n")
-            return
-        
-        print(f"\nJump Lists (Recent {limit}):")
-        print("=" * 140)
-        print("{:<12} {:<20} {:<30} {:<12} {:<20} {:<20} {:<20}".format(
-            "Type", "AppID", "Filename", "Size", "Path", "Modified", "Accessed"))
-        print("=" * 140)
-        
-        for jl in jumplists[:limit]:
-            size_str = f"{jl['Size'] / 1024:.1f} KB"
-            print("{:<12} {:<20} {:<30} {:<12} {:<20} {:<20} {:<20}".format(
-                jl['Type'][:11],
-                jl['AppID'][:19],
-                jl['Filename'][:29],
-                size_str,
-                os.path.dirname(jl['Path'])[:19],
-                jl['Modified'][:19],
-                jl['Accessed'][:19]
-            ))
-        
-        if len(jumplists) > limit:
-            print(f"\n... and {len(jumplists) - limit} more jumplist files")
-        print(f"\nTotal jumplist files: {len(jumplists)}\n")
-    
-    def do_fileless(self, arg):
-        """Advanced fileless malware and in-memory threat detection
-        Usage: fileless [--pid <pid>] [--verbose] [--min-severity <1-4>]"""
-        
-        args = self._parse_args(arg)
-        pid = int(args['--pid']) if '--pid' in args else None
-        verbose = '--verbose' in args
-        min_severity = int(args.get('--min-severity', 1))
-        
-        print("\n[*] Starting advanced fileless malware hunt...")
-        print("[*] Scanning for: PowerShell attacks, credential theft, persistence, RATs, miners...")
-        
-        if pid:
-            detections = FilelessMalwareHunter.scan_process_memory(pid)
-        else:
-            detections = FilelessMalwareHunter.scan_system()
-        
-        # Filter by minimum severity
-        if min_severity > 1:
-            detections = [d for d in detections if FilelessMalwareHunter._get_severity_value(d['severity']) >= min_severity]
-        
-        # Calculate statistics
-        critical_count = sum(1 for d in detections if d['severity'] == 'Critical')
-        high_count = sum(1 for d in detections if d['severity'] == 'High')
-        medium_count = sum(1 for d in detections if d['severity'] == 'Medium')
-        low_count = sum(1 for d in detections if d['severity'] == 'Low')
-        
-        # Determine verdict
-        verdict = "CLEAN"
-        threat_level = "None"
-        
-        if critical_count >= 3:
-            verdict = "CRITICAL INFECTION - IMMEDIATE ACTION REQUIRED"
-            threat_level = "Extreme"
-        elif critical_count > 0:
-            verdict = "CRITICAL THREATS DETECTED"
-            threat_level = "Critical"
-        elif high_count >= 5:
-            verdict = "SEVERE COMPROMISE - HIGH RISK"
-            threat_level = "High"
-        elif high_count > 2:
-            verdict = "MULTIPLE THREATS DETECTED"
-            threat_level = "High"
-        elif high_count > 0:
-            verdict = "THREATS DETECTED"
-            threat_level = "Moderate"
-        elif medium_count > 3:
-            verdict = "SUSPICIOUS ACTIVITY"
-            threat_level = "Low-Medium"
-        elif medium_count > 0:
-            verdict = "POTENTIAL THREATS"
-            threat_level = "Low"
-        
-        # Display results
-        print("\n" + "=" * 80)
-        print("              FILELESS MALWARE HUNTER REPORT")
-        print("=" * 80)
-        
-        print(f"\nVERDICT: {verdict}")
-        print(f"THREAT LEVEL: {threat_level}")
-        
-        print(f"\nSCAN STATISTICS:")
-        print(f"  Total Detections: {len(detections)}")
-        print(f"  Critical: {critical_count}")
-        print(f"  High: {high_count}")
-        print(f"  Medium: {medium_count}")
-        print(f"  Low: {low_count}")
-        
-        if detections:
-            print(f"\nDETAILED FINDINGS:")
-            print("=" * 150)
-            print("{:<25} {:<20} {:<8} {:<30} {:<10} {:<40} {:<20}".format(
-                "Detection Type", "Process", "PID", "Indicator", "Severity", "Address/Details", "MITRE ATT&CK"))
-            print("=" * 150)
-            
-            # Group by severity
-            critical_detections = [d for d in detections if d['severity'] == 'Critical']
-            high_detections = [d for d in detections if d['severity'] == 'High']
-            medium_detections = [d for d in detections if d['severity'] == 'Medium']
-            low_detections = [d for d in detections if d['severity'] == 'Low']
-            
-            # Display critical first
-            for detection in critical_detections + high_detections + medium_detections + low_detections:
-                print("{:<25} {:<20} {:<8} {:<30} {:<10} {:<40} {:<20}".format(
-                    detection['type'][:24],
-                    detection['process'][:19],
-                    detection['pid'],
-                    detection['indicator'][:29],
-                    detection['severity'],
-                    detection['address'][:39],
-                    detection['technique'][:19]
-                ))
-            
-            if critical_count > 0 or high_count > 2:
-                print(f"\n[!] RECOMMENDATION: Immediate incident response required!")
-                print("    Isolate system and conduct full forensic analysis.")
-        else:
-            print(f"\n[+] No fileless malware threats detected.")
-            print("    System appears clean based on advanced memory analysis.")
-        
-        print("\n" + "=" * 80)
-
-    def do_ssdt(self, arg):
-        """List system call table with enhanced live system detection
-        Usage: ssdt [--verbose] [--check-hooks]"""
-        
-        args = self._parse_args(arg)
-        verbose = '--verbose' in args
-        check_hooks = '--check-hooks' in args
-        
-        print("\n[*] Enumerating SSDT entries from live system...")
-        print("[*] Methods: Kernel module analysis, Driver inspection, System call mapping")
-        
-        entries = self.context.get_ssdt_entries()
-        
-        if not entries:
-            print("\nNo SSDT entries found\n")
-            return
-        
-        # Calculate statistics
-        total_entries = len(entries)
-        ntoskrnl_entries = sum(1 for e in entries if 'ntoskrnl' in e.get('Module', '').lower())
-        other_entries = total_entries - ntoskrnl_entries
-        suspicious_count = sum(1 for e in entries if e.get('Status') == 'Suspicious')
-        
-        print(f"\nSystem Service Descriptor Table Analysis:")
-        print("=" * 120)
-        print(f"Total Entries: {total_entries}")
-        print(f"ntoskrnl.exe: {ntoskrnl_entries}")
-        print(f"Other Modules: {other_entries}")
-        print(f"Suspicious: {suspicious_count}")
-        print("=" * 120)
-        
-        if verbose:
-            print("\n{:<8} {:<18} {:<25} {:<35} {:<20} {:<10}".format(
-                "Index", "Address", "Module", "Symbol", "Service Table", "Status"))
-            print("=" * 120)
-            
-            for entry in sorted(entries, key=lambda x: x.get('Index', 0)):
-                status = entry.get('Status', 'Normal')
-                service_table = entry.get('ServiceTable', 'KeServiceDescriptorTable')
-                
-                print("{:<8} {:<18} {:<25} {:<35} {:<20} {:<10}".format(
-                    entry.get('Index', 0),
-                    entry.get('Address', 'N/A')[:17],
-                    entry.get('Module', 'N/A')[:24],
-                    entry.get('Symbol', 'N/A')[:34],
-                    service_table[:19],
-                    status
-                ))
-        else:
-            # Compact view
-            print("\n{:<8} {:<18} {:<25} {:<35} {:<10}".format(
-                "Index", "Address", "Module", "Symbol", "Status"))
-            print("=" * 100)
-            
-            for entry in sorted(entries, key=lambda x: x.get('Index', 0))[:50]:  # Show first 50
-                status = entry.get('Status', 'Normal')
-                
-                print("{:<8} {:<18} {:<25} {:<35} {:<10}".format(
-                    entry.get('Index', 0),
-                    entry.get('Address', 'N/A')[:17],
-                    entry.get('Module', 'N/A')[:24],
-                    entry.get('Symbol', 'N/A')[:34],
-                    status
-                ))
-            
-            if len(entries) > 50:
-                print(f"\n... and {len(entries) - 50} more entries")
-        
-        # Hook detection summary
-        if check_hooks and suspicious_count > 0:
-            print(f"\n[!] SSDT HOOK DETECTION REPORT:")
-            print("=" * 80)
-            suspicious_entries = [e for e in entries if e.get('Status') == 'Suspicious']
-            
-            for entry in suspicious_entries:
-                print(f"  • Index {entry.get('Index')}: {entry.get('Symbol')} -> {entry.get('Module')}")
-                print(f"    Address: {entry.get('Address')}")
-            
-            print(f"\n[!] {suspicious_count} potentially hooked entries detected!")
-            print("    Investigate these modules for rootkit activity.")
-        elif check_hooks:
-            print(f"\n[+] No SSDT hooks detected - SSDT appears clean.")
-        
-        print(f"\nTotal SSDT entries analyzed: {total_entries}")
-        print("Note: Live SSDT analysis provides approximate addresses for demonstration.")
-        print("      For precise SSDT analysis, use kernel debugging tools.\n")
-
-    # EXISTING PLUGIN METHODS
-    def do_pslist(self, arg):
-        """List all running processes
-        Usage: pslist [filter_name]"""
-        
-        processes = self.context.get_processes()
-        filter_name = arg.strip().lower() if arg else None
-        
-        print("\n{:<8} {:<8} {:<25} {:<18} {:<8} {:<8} {:<8} {:<8} {:<20} {:<20}".format(
-            "PID", "PPID", "ImageFileName", "Offset(V)", "Threads", "Handles", "Session", "Wow64", "CreateTime", "ExitTime"))
-        print("=" * 165)
-        
-        for proc in processes:
-            if filter_name and filter_name not in proc['Name'].lower():
-                continue
-            
-            wow64_str = "True" if proc['Wow64'] else "False"
-            start_str = proc['StartTime'].strftime('%Y-%m-%d %H:%M:%S') if proc['StartTime'] else 'N/A'
-            exit_str = proc['ExitTime'].strftime('%Y-%m-%d %H:%M:%S') if proc['ExitTime'] else ''
-            
-            print("{:<8} {:<8} {:<25} {:<18} {:<8} {:<8} {:<8} {:<8} {:<20} {:<20}".format(
-                proc['PID'],
-                proc['PPID'],
-                proc['Name'][:24],
-                proc['Offset'],
-                proc['Threads'],
-                proc['Handles'],
-                proc['SessionId'],
-                wow64_str,
-                start_str,
-                exit_str
-            ))
-        
-        filtered = [p for p in processes if not filter_name or filter_name in p['Name'].lower()]
-        print(f"\nTotal processes: {len(filtered)}\n")
-    
-    def do_psscan(self, arg):
-        """Scan for process structures
-        Usage: psscan [filter_name]"""
-        
-        processes = self.context.get_processes_detailed()
-        filter_name = arg.strip().lower() if arg else None
-        
-        print("\n{:<8} {:<8} {:<25} {:<18} {:<8} {:<8} {:<8} {:<8} {:<20} {:<20}".format(
-            "PID", "PPID", "ImageFileName", "Offset(V)", "Threads", "Handles", "Session", "Wow64", "CreateTime", "ExitTime"))
-        print("=" * 165)
-        
-        for proc in processes:
-            if filter_name and filter_name not in proc['Name'].lower():
-                continue
-            
-            wow64_str = "True" if proc.get('Wow64', False) else "False"
-            start_str = proc['StartTime'].strftime('%Y-%m-%d %H:%M:%S') if proc.get('StartTime') else 'N/A'
-            exit_str = proc['ExitTime'].strftime('%Y-%m-%d %H:%M:%S') if proc.get('ExitTime') else ''
-            
-            print("{:<8} {:<8} {:<25} {:<18} {:<8} {:<8} {:<8} {:<8} {:<20} {:<20}".format(
-                proc['PID'],
-                proc['PPID'],
-                proc['Name'][:24],
-                proc['Offset'],
-                proc['Threads'],
-                proc['Handles'],
-                proc.get('SessionId', 1),
-                wow64_str,
-                start_str,
-                exit_str
-            ))
-        
-        filtered = [p for p in processes if not filter_name or filter_name in p['Name'].lower()]
-        print(f"\nTotal processes found: {len(filtered)}\n")
-    
-    def do_pstree(self, arg):
-        """Display process tree hierarchy
-        Usage: pstree [--verbose]"""
-        
-        processes = self.context.get_processes()
-        proc_dict = {p['PID']: p for p in processes}
-        printed = set()
-        verbose = '--verbose' in arg
-        
-        def print_tree(pid, indent=0, is_last=True):
-            if pid in printed or pid not in proc_dict:
-                return
-            
-            printed.add(pid)
-            proc = proc_dict[pid]
-            
-            if indent == 0:
-                prefix = ""
-                connector = ""
-            else:
-                prefix = "  " * (indent - 1)
-                connector = "└─" if is_last else "├─"
-            
-            if verbose:
-                wow64_str = "W64" if proc['Wow64'] else ""
-                output = f"{prefix}{connector}[{proc['PID']}] {proc['Name']} " \
-                         f"({proc['Offset']}) T:{proc['Threads']} H:{proc['Handles']} " \
-                         f"S:{proc['SessionId']} {wow64_str}"
-            else:
-                output = f"{prefix}{connector}[{proc['PID']}] {proc['Name']}"
-            
-            if len(output) > 140:
-                output = output[:137] + "..."
-            
-            print(output)
-            
-            children = sorted([p for p in processes if p['PPID'] == pid], key=lambda x: x['PID'])
-            for i, child in enumerate(children):
-                is_last_child = (i == len(children) - 1)
-                print_tree(child['PID'], indent + 1, is_last_child)
-        
-        print("\nProcess Tree:")
-        print("=" * 140)
-        
-        root_procs = sorted(
-            [p for p in processes if p['PPID'] == 0 or p['PPID'] not in proc_dict],
-            key=lambda x: x['PID']
-        )
-        
-        for i, proc in enumerate(root_procs):
-            is_last = (i == len(root_procs) - 1)
-            print_tree(proc['PID'], 0, is_last)
-        
-        print()
-    
-    def do_psxview(self, arg):
-        """Cross-view process detection (identifies hidden processes)
-        Usage: psxview"""
-        
-        processes = self.context.get_processes()
-        
-        print("\nProcess Cross-View Analysis:")
-        print("=" * 165)
-        print("{:<8} {:<8} {:<25} {:<18} {:<18} {:<12} {:<12} {:<12} {:<12} {:<12}".format(
-            "PID", "PPID", "ImageFileName", "Offset(V)", "Offset(P)", "pslist", "psscan", "thrdproc", "pspcid", "csrss"))
-        print("=" * 165)
-        
-        for proc in processes:
-            print("{:<8} {:<8} {:<25} {:<18} {:<18} {:<12} {:<12} {:<12} {:<12} {:<12}".format(
-                proc['PID'],
-                proc['PPID'],
-                proc['Name'][:24],
-                proc['Offset'],
-                proc['POffset'],
-                "True",
-                "True",
-                "True",
-                "True",
-                "True"
-            ))
-        
-        print(f"\nTotal processes analyzed: {len(processes)}")
-        print("Note: All processes visible in live analysis appear in all detection methods\n")
-        print("Legend: pslist=Process list enumeration, psscan=Pool scanning,")
-        print("        thrdproc=Thread enumeration, pspcid=CSRSS handles, csrss=CSRSS process list\n")
-    
-    def do_procdup(self, arg):
-        """List processes running multiple instances
-        Usage: procdup [--min <count>]"""
-        
-        args = self._parse_args(arg)
-        min_count = int(args.get('--min', 2))
-        
-        duplicates = self.context.get_duplicate_processes(min_count)
-        
-        if not duplicates:
-            print(f"\nNo processes found with {min_count} or more instances\n")
-            return
-        
-        print(f"\nProcesses with {min_count}+ instances:")
-        print("=" * 120)
-        print("{:<25} {:<8} {:<8} {:<10} {:<60}".format(
-            "Process Name", "PID", "PPID", "Instances", "Path"))
-        print("=" * 120)
-        
-        for dup in duplicates:
-            print("{:<25} {:<8} {:<8} {:<10} {:<60}".format(
-                dup['Name'][:24],
-                dup['PID'],
-                dup['PPID'],
-                dup['Count'],
-                dup['Path'][:59]
-            ))
-        
-        print()
-    
-    def do_proccon(self, arg):
-        """Generate process connection graph (DOT format for Graphviz)
-        Usage: proccon [--output <filepath>]
-        
-        Creates a visual representation of process parent-child relationships.
-        Output is in DOT format which can be visualized with Graphviz tools.
-        
-        Example: proccon --output my_process_tree.dot
-        """
-        
-        args = self._parse_args(arg)
-        output_file = args.get('--output', 'process_tree.dot')
-        
-        print(f"\n[*] Generating process tree visualization...")
-        print(f"[*] Output file: {output_file}")
-        
-        success, proc_count, edge_count, root_count = self.context.generate_dot_file(output_file)
-        
-        if success:
-            print(f"\n[SUCCESS] Process tree graph generated successfully!")
-            print("=" * 80)
-            print(f"  Output File:       {output_file}")
-            print(f"  Total Processes:   {proc_count}")
-            print(f"  Root Processes:    {root_count}")
-            print(f"  Relationships:     {edge_count}")
-            print("=" * 80)
-            print("\nTo visualize the graph, use one of these commands:")
-            print(f"  dot -Tpng {output_file} -o process_tree.png")
-            print(f"  dot -Tsvg {output_file} -o process_tree.svg")
-            print(f"  dot -Tpdf {output_file} -o process_tree.pdf")
-            print("\nNote: Graphviz must be installed to render the visualization.")
-            print("      Download from: https://graphviz.org/download/\n")
-        else:
-            print(f"\n[ERROR] Failed to generate process tree graph.\n")
-    
-    def do_dlllist(self, arg):
-        """List loaded DLLs system-wide or for a process
-        Usage: dlllist [--pid <pid>]"""
-        
-        args = self._parse_args(arg)
-        pid = int(args['--pid']) if '--pid' in args else None
-        
-        try:
-            modules = self.context.get_loaded_modules(pid)
-            
-            if not modules:
-                print(f"\nNo modules found{' for PID ' + str(pid) if pid else ''}\n")
-                return
-            
-            print(f"\nLoaded DLLs{' for PID ' + str(pid) if pid else ' (System-wide)'}:")
-            print("=" * 140)
-            if pid:
-                print("{:<18} {:<70} {:<15}".format("Base", "Path", "Size"))
-            else:
-                print("{:<8} {:<20} {:<18} {:<70} {:<15}".format("PID", "Process", "Base", "Path", "Size"))
-            print("=" * 140)
-            
-            for mod in modules[:100]:
-                size_str = f"{mod['Size'] / 1024:.2f} KB"
-                if pid:
-                    print("{:<18} {:<70} {:<15}".format(mod['Base'], mod['Path'][:69], size_str))
-                else:
-                    print("{:<8} {:<20} {:<18} {:<70} {:<15}".format(
-                        mod['PID'], mod['Process'][:19], mod['Base'], mod['Path'][:69], size_str))
-            
-            if len(modules) > 100:
-                print(f"\n... and {len(modules) - 100} more modules")
-            print(f"\nTotal modules: {len(modules)}\n")
-        except ValueError:
-            print("\nInvalid PID\n")
-    
-    def do_driverscan(self, arg):
-        """Scan for loaded kernel drivers
-        Usage: driverscan [filter_name]"""
-        
-        print("\n[*] Scanning for loaded drivers...")
-        drivers = self.context.get_drivers()
-        
-        if not drivers:
-            print("Unable to retrieve drivers (may require administrator privileges).\n")
-            return
-        
-        filter_name = arg.strip().lower() if arg else None
-        
-        print("\n{:<18} {:<25} {:<35} {:<60}".format("Offset", "Name", "Display Name", "Path"))
-        print("=" * 150)
-        
-        for drv in sorted(drivers, key=lambda x: x['Name']):
-            if filter_name and filter_name not in drv['Name'].lower() and filter_name not in drv['DisplayName'].lower():
-                continue
-            
-            status_map = {1: 'STOP', 2: 'STRT', 3: 'STOP', 4: 'RUN'}
-            status = status_map.get(drv['Status'], 'UNK')
-            display = f"{drv['DisplayName'][:30]} [{status}]"
-            
-            print("{:<18} {:<25} {:<35} {:<60}".format(
-                drv['Offset'],
-                drv['Name'][:24],
-                display,
-                drv['Path'][:59]
-            ))
-        
-        filtered = [d for d in drivers if not filter_name or 
-                   filter_name in d['Name'].lower() or 
-                   filter_name in d['DisplayName'].lower()]
-        print(f"\nTotal drivers: {len(filtered)}\n")
-    
-    def do_netscan(self, arg):
-        """Display network connections
-        Usage: netscan [--pid <pid>]"""
-        
-        args = self._parse_args(arg)
-        connections = self.context.get_network_connections()
-        filter_pid = None
-        
-        if '--pid' in args:
-            try:
-                filter_pid = int(args['--pid'])
-            except ValueError:
-                print("\nInvalid PID\n")
-                return
-        
-        print("\n{:<8} {:<30} {:<30} {:<15} {:<10}".format(
-            "PID", "Local Address", "Remote Address", "State", "Proto"))
-        print("=" * 110)
-        
-        for conn in connections:
-            if filter_pid and conn['PID'] != filter_pid:
-                continue
-            
-            print("{:<8} {:<30} {:<30} {:<15} {:<10}".format(
-                conn['PID'],
-                conn['LocalAddr'][:29],
-                conn['RemoteAddr'][:29],
-                conn['Status'][:14],
-                conn['Type'][:9]
-            ))
-        
-        print()
-    
-    def do_netstat(self, arg):
-        """Alias for netscan"""
-        self.do_netscan(arg)
-    
-    def do_handles(self, arg):
-        """List handles system-wide or for a process
-        Usage: handles [--pid <pid>]"""
-        
-        args = self._parse_args(arg)
-        pid = int(args['--pid']) if '--pid' in args else None
-        
-        try:
-            handles = self.context.get_handles(pid)
-            
-            if not handles:
-                print(f"\nNo handles found{' for PID ' + str(pid) if pid else ''}\n")
-                return
-            
-            print(f"\nHandles{' for PID ' + str(pid) if pid else ' (System-wide)'}:")
-            print("=" * 140)
-            if pid:
-                print("{:<15} {:<120}".format("Type", "Details"))
-            else:
-                print("{:<8} {:<20} {:<15} {:<90}".format("PID", "Process", "Type", "Details"))
-            print("=" * 140)
-            
-            for handle in handles[:100]:
-                if handle['Type'] == 'File':
-                    details = handle['Path']
-                elif handle['Type'] == 'Network':
-                    details = f"{handle.get('LocalAddr', 'N/A')} -> {handle.get('RemoteAddr', 'N/A')}"
-                else:
-                    details = "N/A"
-                
-                if pid:
-                    print("{:<15} {:<120}".format(handle['Type'], details[:119]))
-                else:
-                    print("{:<8} {:<20} {:<15} {:<90}".format(
-                        handle['PID'], handle['Process'][:19], handle['Type'], details[:89]))
-            
-            if len(handles) > 100:
-                print(f"\n... and {len(handles) - 100} more handles")
-            print(f"\nTotal handles: {len(handles)}\n")
-        except ValueError:
-            print("\nInvalid PID\n")
-    
-    def do_cmdline(self, arg):
-        """Display command line for processes
-        Usage: cmdline [--pid <pid>]"""
-        
-        args = self._parse_args(arg)
-        processes = self.context.get_processes()
-        filter_pid = None
-        
-        if '--pid' in args:
-            try:
-                filter_pid = int(args['--pid'])
-            except ValueError:
-                print("\nInvalid PID\n")
-                return
-        
-        print("\n{:<8} {:<25} {:<100}".format("PID", "Process", "Command Line"))
-        print("=" * 140)
-        
-        for proc in processes:
-            if filter_pid and proc['PID'] != filter_pid:
-                continue
-            
-            if proc['CommandLine'] != 'N/A':
-                print("{:<8} {:<25} {:<100}".format(
-                    proc['PID'],
-                    proc['Name'][:24],
-                    proc['CommandLine'][:99]
-                ))
-        
-        print()
-    
-    def do_envars(self, arg):
-        """Display environment variables system-wide or for a process
-        Usage: envars [--pid <pid>]"""
-        
-        args = self._parse_args(arg)
-        pid = int(args['--pid']) if '--pid' in args else None
-        
-        try:
-            env_data = self.context.get_environment_variables(pid)
-            
-            if not env_data:
-                print(f"\nNo environment variables found{' for PID ' + str(pid) if pid else ''}\n")
-                return
-            
-            print(f"\nEnvironment Variables{' for PID ' + str(pid) if pid else ' (System-wide)'}:")
-            print("=" * 140)
-            if pid:
-                print("{:<30} {:<105}".format("Variable", "Value"))
-            else:
-                print("{:<8} {:<20} {:<30} {:<75}".format("PID", "Process", "Variable", "Value"))
-            print("=" * 140)
-            
-            for item in env_data[:100]:
-                if pid:
-                    print("{:<30} {:<105}".format(item['Variable'][:29], item['Value'][:104]))
-                else:
-                    print("{:<8} {:<20} {:<30} {:<75}".format(
-                        item['PID'], item['Process'][:19], item['Variable'][:29], item['Value'][:74]))
-            
-            if len(env_data) > 100:
-                print(f"\n... and {len(env_data) - 100} more variables")
-            print(f"\nTotal variables: {len(env_data)}\n")
-        except ValueError:
-            print("\nInvalid PID\n")
-    
-    def do_threads(self, arg):
-        """List threads for a process or all threads
-        Usage: threads [--pid <pid>]"""
-        
-        args = self._parse_args(arg)
-        pid = None
-        
-        if '--pid' in args:
-            try:
-                pid = int(args['--pid'])
-            except ValueError:
-                print("\nInvalid PID\n")
-                return
-        
-        threads = self.context.get_threads(pid)
-        
-        if not threads:
-            print(f"\nNo threads found{' for PID ' + str(pid) if pid else ''}\n")
-            return
-        
-        print(f"\nThreads{' for PID ' + str(pid) if pid else ''}:")
-        print("=" * 100)
-        print("{:<8} {:<12} {:<15} {:<15}".format("PID", "TID", "UserTime", "SystemTime"))
-        print("=" * 100)
-        
-        for thread in threads[:50]:
-            print("{:<8} {:<12} {:<15.2f} {:<15.2f}".format(
-                thread['PID'],
-                thread['TID'],
-                thread['UserTime'],
-                thread['SystemTime']
-            ))
-        
-        if len(threads) > 50:
-            print(f"\n... and {len(threads) - 50} more threads")
-        print(f"\nTotal threads: {len(threads)}\n")
-    
-    def do_privileges(self, arg):
-        """List process privileges
-        Usage: privileges --pid <pid>"""
-        
-        args = self._parse_args(arg)
-        
-        if '--pid' not in args:
-            print("\nUsage: privileges --pid <pid>\n")
-            return
-        
-        try:
-            pid = int(args['--pid'])
-        except ValueError:
-            print("\nInvalid PID\n")
-            return
-        
-        privileges = self.context.get_privileges(pid)
-        
-        print(f"\nPrivileges for PID {pid}:")
-        print("=" * 80)
-        print("{:<40} {:<30}".format("Privilege", "Attributes"))
-        print("=" * 80)
-        
-        for priv in privileges:
-            print("{:<40} {:<30}".format(priv['Privilege'], priv['Attributes']))
-        
-        print()
-    
-    def do_getsids(self, arg):
-        """Display process SIDs
-        Usage: getsids [--pid <pid>]"""
-        
-        args = self._parse_args(arg)
-        pid = None
-        
-        if '--pid' in args:
-            try:
-                pid = int(args['--pid'])
-            except ValueError:
-                print("\nInvalid PID\n")
-                return
-        
-        print(f"\nProcess SIDs{' for PID ' + str(pid) if pid else ''}:")
-        print("=" * 100)
-        print("{:<8} {:<30} {:<50}".format("PID", "Process", "SID/Username"))
-        print("=" * 100)
-        
-        try:
-            sids = self.context.get_sids(pid)
-            
-            for sid in sids:
-                try:
-                    print("{:<8} {:<30} {:<50}".format(
-                        sid.get('PID', 0),
-                        sid.get('Process', 'Unknown')[:29],
-                        sid.get('SID', 'N/A')[:49]
-                    ))
-                except Exception:
-                    continue
-            
-            if not sids:
-                print("No SIDs found")
-            print()
-        except Exception as e:
-            print(f"\nError: {e}\n")
-    
-    def do_services(self, arg):
-        """List Windows services
-        Usage: services [--filter <name>]"""
-        
-        args = self._parse_args(arg)
-        services = self.context.get_services()
-        
-        if not services:
-            print("\nUnable to retrieve services (may require administrator privileges)\n")
-            return
-        
-        filter_name = args.get('--filter', '').lower()
-        if filter_name:
-            services = [s for s in services if filter_name in s['Name'].lower() or 
-                       filter_name in s['DisplayName'].lower()]
-        
-        print("\n{:<30} {:<50} {:<10}".format("Service Name", "Display Name", "Status"))
-        print("=" * 100)
-        
-        for svc in services:
-            status_map = {1: 'STOPPED', 2: 'START', 3: 'STOP', 4: 'RUNNING'}
-            status_str = status_map.get(svc['Status'], 'UNKNOWN')
-            
-            print("{:<30} {:<50} {:<10}".format(
-                svc['Name'][:29],
-                svc['DisplayName'][:49],
-                status_str
-            ))
-        
-        print(f"\nTotal services: {len(services)}\n")
-    
-    def do_getservicesids(self, arg):
-        """Display service SIDs from registry
-        Usage: getservicesids [--filter <name>]"""
-        
-        args = self._parse_args(arg)
-        filter_pattern = args.get('--filter', '').lower()
-        
-        print("\n[*] Retrieving service SIDs from registry...")
-        service_sids = self.context.get_service_sids()
-        
-        if not service_sids:
-            print("\nUnable to retrieve service SIDs (may require administrator privileges)\n")
-            return
-        
-        if filter_pattern:
-            service_sids = [s for s in service_sids if filter_pattern in s['Service'].lower() or 
-                          filter_pattern in s['SID'].lower()]
-        
-        print("\nService SIDs:")
-        print("=" * 120)
-        print("{:<70} {:<45}".format("SID", "Service"))
-        print("=" * 120)
-        
-        for svc in service_sids:
-            print("{:<70} {:<45}".format(
-                svc['SID'][:69],
-                svc['Service'][:44]
-            ))
-        
-        print(f"\nTotal service SIDs: {len(service_sids)}\n")
-    
-    def do_svcscan(self, arg):
-        """Alias for services"""
-        self.do_services(arg)
-    
-    def do_svclist(self, arg):
-        """Alias for services"""
-        self.do_services(arg)
-    
-    def do_modules(self, arg):
-        """List all loaded modules system-wide
-        Usage: modules [--filter <pattern>]"""
-        
-        args = self._parse_args(arg)
-        filter_pattern = args.get('--filter', '').lower()
-        
-        print("\n[*] Scanning for loaded modules system-wide...")
-        modules = self.context.get_modules()
-        
-        if filter_pattern:
-            modules = [m for m in modules if filter_pattern in m['Name'].lower() or 
-                      filter_pattern in m['Path'].lower()]
-        
-        print("\nLoaded Modules (System-wide):")
-        print("=" * 140)
-        print("{:<40} {:<20} {:<15} {:<60}".format("Module Name", "Base Address", "Size", "Path"))
-        print("=" * 140)
-        
-        for mod in modules[:100]:
-            print("{:<40} {:<20} {:<15} {:<60}".format(
-                mod['Name'][:39],
-                mod['Base'],
-                mod['Size'],
-                mod['Path'][:59]
-            ))
-        
-        if len(modules) > 100:
-            print(f"\n... and {len(modules) - 100} more modules")
-        print(f"\nTotal modules: {len(modules)}\n")
-    
-    def do_modscan(self, arg):
-        """Scan for kernel modules and DLLs system-wide
-        Usage: modscan [--filter <pattern>]"""
-        
-        args = self._parse_args(arg)
-        filter_pattern = args.get('--filter', '').lower()
-        
-        print("\n[*] Scanning for kernel modules and DLLs...")
-        modules = self.context.get_modules()
-        
-        if filter_pattern:
-            modules = [m for m in modules if filter_pattern in m['Name'].lower() or 
-                      filter_pattern in m['Path'].lower()]
-        
-        print("\nModule Scan Results:")
-        print("=" * 140)
-        print("{:<40} {:<20} {:<15} {:<60}".format("Module Name", "Base Address", "Size", "Path"))
-        print("=" * 140)
-        
-        for mod in modules[:100]:
-            print("{:<40} {:<20} {:<15} {:<60}".format(
-                mod['Name'][:39],
-                mod['Base'],
-                mod['Size'],
-                mod['Path'][:59]
-            ))
-        
-        if len(modules) > 100:
-            print(f"\n... and {len(modules) - 100} more modules")
-        print(f"\nTotal modules found: {len(modules)}\n")
-    
-    def do_callbacks(self, arg):
-        """List kernel callbacks
-        Usage: callbacks"""
-        
-        print("\n[*] Enumerating kernel callbacks...")
-        callbacks = self.context.get_callbacks()
-        
-        if not callbacks:
-            print("\nNo callbacks found (may require administrator privileges)\n")
-            return
-        
-        print("\nKernel Callbacks:")
-        print("=" * 120)
-        print("{:<30} {:<20} {:<30} {:<35}".format("Type", "Address", "Module", "Detail"))
-        print("=" * 120)
-        
-        for cb in callbacks[:50]:
-            print("{:<30} {:<20} {:<30} {:<35}".format(
-                cb['Type'][:29],
-                cb['Address'][:19],
-                cb['Module'][:29],
-                cb['Detail'][:34]
-            ))
-        
-        if len(callbacks) > 50:
-            print(f"\n... and {len(callbacks) - 50} more callbacks")
-        print(f"\nTotal callbacks: {len(callbacks)}\n")
-    
-    def do_timers(self, arg):
-        """List kernel timers (system thread timing information)
-        Usage: timers [--limit <n>]"""
-        
-        args = self._parse_args(arg)
-        limit = int(args.get('--limit', 50))
-        
-        print("\n[*] Enumerating kernel timers...")
-        timers = self.context.get_kernel_timers()
-        
-        if not timers:
-            print("\nNo kernel timers found\n")
-            return
-        
-        print("\nKernel Timers:")
-        print("=" * 120)
-        print("{:<20} {:<15} {:<15} {:<15} {:<25} {:<25}".format(
-            "Offset", "DueTime", "Period", "Signaled", "Module", "Type"))
-        print("=" * 120)
-        
-        for timer in timers[:limit]:
-            print("{:<20} {:<15} {:<15} {:<15} {:<25} {:<25}".format(
-                timer['Offset'][:19],
-                timer['DueTime'][:14],
-                timer['Period'][:14],
-                timer['Signaled'][:14],
-                timer['Module'][:24],
-                timer['Type'][:24]
-            ))
-        
-        if len(timers) > limit:
-            print(f"\n... and {len(timers) - limit} more timers")
-        print(f"\nTotal timers: {len(timers)}\n")
-    
-    def do_consoles(self, arg):
-        """Display console command history
-        Usage: consoles"""
-        
-        print("\n[*] Retrieving console command history...")
-        history = self.context.get_console_history()
-        
-        if not history:
-            print("\nNo console history found\n")
-            return
-        
-        print("\nConsole Command History:")
-        print("=" * 120)
-        print("{:<8} {:<20} {:<15} {:<70}".format("Index", "Application", "Source", "Command"))
-        print("=" * 120)
-        
-        for entry in history[:100]:
-            print("{:<8} {:<20} {:<15} {:<70}".format(
-                entry['Index'],
-                entry['Application'][:19],
-                entry['Source'][:14],
-                entry['Command'][:69]
-            ))
-        
-        if len(history) > 100:
-            print(f"\n... and {len(history) - 100} more commands")
-        print(f"\nTotal commands: {len(history)}\n")
-    
-    def do_cmdscan(self, arg):
-        """Alias for consoles - Display command history
-        Usage: cmdscan"""
-        self.do_consoles(arg)
-    
-    def do_desktops(self, arg):
-        """Display files and folders on the desktop
-        Usage: desktops [--limit <n>]"""
-        
-        args = self._parse_args(arg)
-        limit = int(args.get('--limit', 100))
-        
-        print("\n[*] Enumerating desktop files and folders...")
-        desktop_items = self.context.get_desktop_files()
-        
-        if not desktop_items:
-            print("\nNo desktop items found\n")
-            return
-        
-        import datetime
-        
-        print("\nDesktop Files and Folders:")
-        print("=" * 140)
-        print("{:<50} {:<10} {:<15} {:<20} {:<10} {:<30}".format(
-            "Name", "Type", "Size", "Modified", "Location", "Path"))
-        print("=" * 140)
-        
-        for item in desktop_items[:limit]:
-            size_str = f"{item['Size'] / 1024:.2f} KB" if item['Size'] > 0 else "N/A"
-            try:
-                mod_time = datetime.datetime.fromtimestamp(item['Modified']).strftime('%Y-%m-%d %H:%M:%S')
-            except:
-                mod_time = "N/A"
-            
-            print("{:<50} {:<10} {:<15} {:<20} {:<10} {:<30}".format(
-                item['Name'][:49],
-                item['Type'][:9],
-                size_str[:14],
-                mod_time[:19],
-                item['Location'][:9],
-                item['Path'][:29]
-            ))
-        
-        if len(desktop_items) > limit:
-            print(f"\n... and {len(desktop_items) - limit} more items")
-        print(f"\nTotal desktop items: {len(desktop_items)}\n")
-    
-    def do_deskscan(self, arg):
-        """Enumerate desktop windows and window stations
-        Usage: deskscan [--limit <n>]"""
-        
-        args = self._parse_args(arg)
-        limit = int(args.get('--limit', 50))
-        
-        print("\n[*] Enumerating desktop windows...")
-        desktops = self.context.get_desktops()
-        
-        if not desktops:
-            print("\nNo desktop windows found\n")
-            return
-        
-        print("\nDesktop Windows:")
-        print("=" * 140)
-        print("{:<18} {:<15} {:<10} {:<15} {:<25} {:<8} {:<50}".format(
-            "Offset", "WindowStation", "Session", "Desktop", "Process", "PID", "Title"))
-        print("=" * 140)
-        
-        for desktop in desktops[:limit]:
-            print("{:<18} {:<15} {:<10} {:<15} {:<25} {:<8} {:<50}".format(
-                desktop['Offset'][:17],
-                desktop['WindowStation'][:14],
-                desktop['Session'],
-                desktop['Desktop'][:14],
-                desktop['Process'][:24],
-                desktop['PID'],
-                desktop.get('Title', 'N/A')[:49]
-            ))
-        
-        if len(desktops) > limit:
-            print(f"\n... and {len(desktops) - limit} more windows")
-        print(f"\nTotal windows: {len(desktops)}\n")
-    
-    def do_joblinks(self, arg):
-        """Display process job object relationships
-        Usage: joblinks [--limit <n>]"""
-        
-        args = self._parse_args(arg)
-        limit = int(args.get('--limit', 100))
-        
-        print("\n[*] Analyzing process job relationships...")
-        job_links = self.context.get_job_links()
-        
-        if not job_links:
-            print("\nNo job links found\n")
-            return
-        
-        print("\nProcess Job Links:")
-        print("=" * 150)
-        print("{:<18} {:<25} {:<8} {:<8} {:<8} {:<10} {:<8} {:<8} {:<12} {:<50}".format(
-            "Offset", "Name", "PID", "PPID", "Session", "TotalProcs", "Active", "Terminated", "JobLink", "Path"))
-        print("=" * 150)
-        
-        for job in job_links[:limit]:
-            print("{:<18} {:<25} {:<8} {:<8} {:<8} {:<10} {:<8} {:<8} {:<12} {:<50}".format(
-                job['Offset'][:17],
-                job['Name'][:24],
-                job['PID'],
-                job['PPID'],
-                job['Session'],
-                job['TotalProcs'],
-                job['ActiveProcs'],
-                job['TerminatedProcs'],
-                job['JobLink'][:11],
-                job['Path'][:49] if job['Path'] != 'N/A' else 'N/A'
-            ))
-        
-        if len(job_links) > limit:
-            print(f"\n... and {len(job_links) - limit} more job links")
-        print(f"\nTotal job links: {len(job_links)}\n")
-    
-    def do_scheduledtasks(self, arg):
-        """Display scheduled tasks (detailed)
-        Usage: scheduledtasks [--limit <n>]"""
-        
-        args = self._parse_args(arg)
-        limit = int(args.get('--limit', 100))
-        
-        print("\n[*] Retrieving scheduled tasks...")
-        tasks = self.context.get_scheduled_tasks_registry()
-        
-        if not tasks:
-            print("\nNo scheduled tasks found\n")
-            return
-        
-        print("\nScheduled Tasks:")
-        print("=" * 160)
-        print("{:<40} {:<15} {:<10} {:<25} {:<25} {:<40}".format(
-            "Task Name", "State", "Enabled", "Last Run", "Next Run", "Actions"))
-        print("=" * 160)
-        
-        for task in tasks[:limit]:
-            enabled_str = "Yes" if task['Enabled'] else "No"
-            print("{:<40} {:<15} {:<10} {:<25} {:<25} {:<40}".format(
-                task['Name'][:39],
-                task['State'][:14],
-                enabled_str,
-                task['LastRun'][:24],
-                task['NextRun'][:24],
-                task['Actions'][:39]
-            ))
-        
-        if len(tasks) > limit:
-            print(f"\n... and {len(tasks) - limit} more tasks")
-        print(f"\nTotal scheduled tasks: {len(tasks)}\n")
-    
-    def do_deletedfiles(self, arg):
-        """Display recently deleted files from Recycle Bin and registry
-        Usage: deletedfiles [--type <ext>] [--limit <n>]"""
-        
-        args = self._parse_args(arg)
-        file_type = args.get('--type', '').lower()
-        limit = int(args.get('--limit', 100))
-        
-        print("\n[*] Scanning for deleted files (Recycle Bin + Registry)...")
-        deleted = self.context.get_deleted_files()
-        
-        if file_type:
-            deleted = [f for f in deleted if file_type in f['Type'].lower()]
-        
-        if not deleted:
-            print(f"\nNo deleted files found{' matching type filter' if file_type else ''}\n")
-            return
-        
-        print("\nDeleted Files (Recycle Bin + Registry):")
-        print("=" * 150)
-        print("{:<50} {:<15} {:<25} {:<15} {:<40}".format(
-            "Original Name", "Type", "Date Deleted", "Size", "Location/Path"))
-        print("=" * 150)
-        
-        for item in deleted[:limit]:
-            size_str = f"{item['Size'] / 1024:.2f} KB" if isinstance(item['Size'], (int, float)) and item['Size'] > 0 else "N/A"
-            
-            print("{:<50} {:<15} {:<25} {:<15} {:<40}".format(
-                item['OriginalName'][:49],
-                item['Type'][:14],
-                item['DateDeleted'][:24],
-                size_str[:14],
-                item['DeletedPath'][:39]
-            ))
-        
-        if len(deleted) > limit:
-            print(f"\n... and {len(deleted) - limit} more deleted files")
-        print(f"\nTotal deleted files: {len(deleted)}\n")
-    
-    def do_recentdocs(self, arg):
-        """Display recently opened documents from filesystem and registry
-        Usage: recentdocs [--type <ext>] [--limit <n>]"""
-        
-        args = self._parse_args(arg)
-        doc_type = args.get('--type', '').lower()
-        limit = int(args.get('--limit', 100))
-        
-        print("\n[*] Retrieving recently opened documents (Filesystem + Registry)...")
-        recent = self.context.get_recent_documents()
-        
-        if doc_type:
-            recent = [d for d in recent if doc_type in d['Type'].lower()]
-        
-        if not recent:
-            print(f"\nNo recent documents found{' matching type filter' if doc_type else ''}\n")
-            return
-        
-        print("\nRecently Opened Documents:")
-        print("=" * 155)
-        print("{:<50} {:<15} {:<25} {:<25} {:<35}".format(
-            "Document Name", "Type", "Last Accessed", "Modified", "Target Path"))
-        print("=" * 155)
-        
-        for doc in recent[:limit]:
-            print("{:<50} {:<15} {:<25} {:<25} {:<35}".format(
-                doc['DocumentName'][:49],
-                doc['Type'][:14],
-                doc['LastAccessed'][:24],
-                doc['Modified'][:24],
-                doc['TargetPath'][:34]
-            ))
-        
-        if len(recent) > limit:
-            print(f"\n... and {len(recent) - limit} more documents")
-        print(f"\nTotal recent documents: {len(recent)}\n")
-    
-    def do_userassist(self, arg):
-        """Display UserAssist data - program execution tracking
-        Usage: userassist [--limit <n>] [--filter <name>]"""
-        
-        args = self._parse_args(arg)
-        limit = int(args.get('--limit', 100))
-        filter_name = args.get('--filter', '').lower()
-        
-        print("\n[*] Retrieving UserAssist data (program execution tracking)...")
-        userassist = self.context.get_userassist()
-        
-        if filter_name:
-            userassist = [u for u in userassist if filter_name in u['Program'].lower()]
-        
-        if not userassist:
-            print(f"\nNo UserAssist data found{' matching filter' if filter_name else ''}\n")
-            return
-        
-        print("\nUserAssist - Program Execution Tracking:")
-        print("=" * 155)
-        print("{:<80} {:<12} {:<25} {:<35}".format(
-            "Program/Path", "Run Count", "Last Executed", "Type"))
-        print("=" * 155)
-        
-        for item in userassist[:limit]:
-            print("{:<80} {:<12} {:<25} {:<35}".format(
-                item['Program'][:79],
-                str(item['RunCount']),
-                item['LastExecuted'][:24],
-                item['Type'][:34]
-            ))
-        
-        if len(userassist) > limit:
-            print(f"\n... and {len(userassist) - limit} more entries")
-        print(f"\nTotal UserAssist entries: {len(userassist)}\n")
-    
-    def do_persistence(self, arg):
-        """Display persistence mechanisms - startup programs, scheduled tasks, suspicious executables
-        Usage: persistence [--type <startup|registry|exe|task>] [--limit <n>]"""
-        
-        args = self._parse_args(arg)
-        limit = int(args.get('--limit', 150))
-        filter_type = args.get('--type', '').lower()
-        
-        print("\n[*] Scanning for persistence mechanisms...")
-        print("[*] Checking: Startup folders, Registry Run keys, Scheduled Tasks, TEMP/APPDATA executables...")
-        persistence = self.context.get_persistence_mechanisms()
-        
-        if filter_type:
-            type_mapping = {
-                'startup': 'Startup Folder',
-                'registry': 'Registry Run Key',
-                'exe': 'Suspicious EXE',
-                'task': 'Scheduled Task'
-            }
-            if filter_type in type_mapping:
-                persistence = [p for p in persistence if p['Type'] == type_mapping[filter_type]]
-        
-        if not persistence:
-            print(f"\nNo persistence mechanisms found{' matching filter' if filter_type else ''}\n")
-            return
-        
-        print("\nPersistence Mechanisms:")
-        print("=" * 165)
-        print("{:<35} {:<25} {:<20} {:<60} {:<20}".format(
-            "Name", "Location", "Type", "Target/Path", "Created"))
-        print("=" * 165)
-        
-        for item in persistence[:limit]:
-            print("{:<35} {:<25} {:<20} {:<60} {:<20}".format(
-                item['Name'][:34],
-                item['Location'][:24],
-                item['Type'][:19],
-                item['Target'][:59],
-                item['Created'][:19]
-            ))
-        
-        if len(persistence) > limit:
-            print(f"\n... and {len(persistence) - limit} more items")
-        print(f"\nTotal persistence mechanisms: {len(persistence)}\n")
-        print("TIP: Use --type to filter: startup, registry, exe, task\n")
-    
-    def do_info(self, arg):
-        """Display system information
-        Usage: info"""
-        
-        import platform
-        
-        print("\nSystem Information:")
-        print("=" * 80)
-        print(f"OS: {platform.system()} {platform.release()}")
-        print(f"Version: {platform.version()}")
-        print(f"Architecture: {platform.machine()}")
-        print(f"Processor: {platform.processor()}")
-        print(f"Hostname: {platform.node()}")
-        
-        mem = psutil.virtual_memory()
-        print(f"\nMemory:")
-        print(f"  Total: {mem.total / (1024**3):.2f} GB")
-        print(f"  Available: {mem.available / (1024**3):.2f} GB")
-        print(f"  Used: {mem.used / (1024**3):.2f} GB ({mem.percent}%)")
-        
-        print(f"\nCPU:")
-        print(f"  Physical cores: {psutil.cpu_count(logical=False)}")
-        print(f"  Logical cores: {psutil.cpu_count(logical=True)}")
-        print(f"  Usage: {psutil.cpu_percent(interval=1)}%")
-        
-        print()
-    
-    def do_filescan(self, arg):
-        """Scan for file objects
-        Usage: filescan [--filter <pattern>]"""
-        
-        args = self._parse_args(arg)
-        filter_pattern = args.get('--filter', '').lower()
-        
-        print("\n[*] Scanning for file handles across all processes...")
-        
-        files = []
-        for proc in psutil.process_iter(['pid', 'name']):
-            try:
-                for oh in proc.open_files():
-                    if not filter_pattern or filter_pattern in oh.path.lower():
-                        files.append({
-                            'PID': proc.info['pid'],
-                            'Process': proc.info['name'],
-                            'Path': oh.path
-                        })
-            except (psutil.AccessDenied, psutil.NoSuchProcess):
-                continue
-        
-        print("\n{:<8} {:<25} {:<80}".format("PID", "Process", "File Path"))
-        print("=" * 120)
-        
-        for f in files[:50]:
-            print("{:<8} {:<25} {:<80}".format(
-                f['PID'],
-                f['Process'][:24],
-                f['Path'][:79]
-            ))
-        
-        if len(files) > 50:
-            print(f"\n... and {len(files) - 50} more files")
-        print(f"\nTotal files: {len(files)}\n")
-    
-    def do_vadinfo(self, arg):
-        """Display VAD information system-wide or for a process
-        Usage: vadinfo [--pid <pid>]"""
-        
-        args = self._parse_args(arg)
-        pid = int(args['--pid']) if '--pid' in args else None
-        
-        try:
-            vad_data = self.context.get_vad_info(pid)
-            
-            if not vad_data:
-                print(f"\nNo VAD information found{' for PID ' + str(pid) if pid else ''}\n")
-                return
-            
-            print(f"\nVAD Information{' for PID ' + str(pid) if pid else ' (System-wide)'}:")
-            print("=" * 140)
-            if pid:
-                print("{:<20} {:<15} {:<100}".format("Address", "Size", "Path"))
-            else:
-                print("{:<8} {:<20} {:<20} {:<15} {:<70}".format("PID", "Process", "Address", "Size", "Path"))
-            print("=" * 140)
-            
-            for vad in vad_data[:100]:
-                size_str = f"{vad['Size'] / 1024:.2f} KB"
-                if pid:
-                    print("{:<20} {:<15} {:<100}".format(vad['Address'], size_str, vad['Path'][:99]))
-                else:
-                    print("{:<8} {:<20} {:<20} {:<15} {:<70}".format(
-                        vad['PID'], vad['Process'][:19], vad['Address'], size_str, vad['Path'][:69]))
-            
-            if len(vad_data) > 100:
-                print(f"\n... and {len(vad_data) - 100} more VAD nodes")
-            print(f"\nTotal VAD nodes: {len(vad_data)}\n")
-        except ValueError:
-            print("\nInvalid PID\n")
-    
-    def do_memmap(self, arg):
-        """Display memory map system-wide or for a process
-        Usage: memmap [--pid <pid>]"""
-        
-        args = self._parse_args(arg)
-        pid = int(args['--pid']) if '--pid' in args else None
-        
-        try:
-            vad_data = self.context.get_vad_info(pid)
-            
-            if not vad_data:
-                print(f"\nNo memory map found{' for PID ' + str(pid) if pid else ''}\n")
-                return
-            
-            print(f"\nMemory Map{' for PID ' + str(pid) if pid else ' (System-wide)'}:")
-            print("=" * 140)
-            if pid:
-                print("{:<20} {:<15} {:<15} {:<85}".format("Virtual Address", "Size", "RSS", "Path"))
-            else:
-                print("{:<8} {:<20} {:<20} {:<15} {:<70}".format("PID", "Process", "Virtual Address", "Size", "Path"))
-            print("=" * 140)
-            
-            for vad in vad_data[:100]:
-                size_str = f"{vad['Size'] / 1024:.2f} KB"
-                if pid:
-                    print("{:<20} {:<15} {:<15} {:<85}".format(
-                        vad['Address'], size_str, size_str, vad['Path'][:84]))
-                else:
-                    print("{:<8} {:<20} {:<20} {:<15} {:<70}".format(
-                        vad['PID'], vad['Process'][:19], vad['Address'], size_str, vad['Path'][:69]))
-            
-            if len(vad_data) > 100:
-                print(f"\n... and {len(vad_data) - 100} more memory regions")
-            print(f"\nTotal memory regions: {len(vad_data)}\n")
-        except ValueError:
-            print("\nInvalid PID\n")
-    
-    def do_search(self, arg):
-        """Search for processes by name or PID
-        Usage: search <term>"""
-        
-        if not arg.strip():
-            print("\nUsage: search <term>\n")
-            return
-        
-        term = arg.strip().lower()
-        processes = self.context.get_processes()
-        
-        results = []
-        for proc in processes:
-            if (term in proc['Name'].lower() or 
-                term in str(proc['PID']) or 
-                (proc['Path'] != 'N/A' and term in proc['Path'].lower()) or
-                (proc['CommandLine'] != 'N/A' and term in proc['CommandLine'].lower())):
-                results.append(proc)
-        
-        if not results:
-            print(f"\nNo processes found matching '{term}'\n")
-            return
-        
-        print(f"\nSearch Results for '{term}':")
-        print("=" * 120)
-        print("{:<8} {:<25} {:<8} {:<50}".format("PID", "Process Name", "PPID", "Path"))
-        print("=" * 120)
-        
-        for proc in results:
-            print("{:<8} {:<25} {:<8} {:<50}".format(
-                proc['PID'],
-                proc['Name'][:24],
-                proc['PPID'],
-                proc['Path'][:49] if proc['Path'] != 'N/A' else 'N/A'
-            ))
-        
-        print(f"\nFound {len(results)} matching processes\n")
-    
-    def do_plugins(self, arg):
-        """List all available plugins"""
-        
-        plugins = [
-            ('pslist', 'List running processes'),
-            ('psscan', 'Scan for process structures'),
-            ('pstree', 'Display process tree'),
-            ('psxview', 'Cross-view process detection'),
-            ('procdup', 'List duplicate process instances'),
-            ('proccon', 'Generate process connection graph (DOT/Graphviz)'),
-            ('cmdline', 'Display process command lines'),
-            ('dlllist', 'List loaded DLLs'),
-            ('handles', 'List process handles'),
-            ('memmap', 'Display memory map'),
-            ('vadinfo', 'Display VAD information'),
-            ('netscan', 'Scan network connections'),
-            ('netstat', 'Display network statistics'),
-            ('modules', 'List all modules system-wide'),
-            ('modscan', 'Scan for modules system-wide'),
-            ('driverscan', 'Scan for drivers'),
-            ('services', 'List Windows services'),
-            ('getservicesids', 'Display service SIDs'),
-            ('svcscan', 'Scan for services'),
-            ('svclist', 'List services'),
-            ('callbacks', 'List kernel callbacks'),
-            ('ssdt', 'List system call table'),
-            ('timers', 'List kernel timers'),
-            ('consoles', 'Display console command history'),
-            ('cmdscan', 'Scan command history'),
-            ('desktops', 'Display desktop files and folders'),
-            ('deskscan', 'Scan desktop windows'),
-            ('joblinks', 'Display process job relationships'),
-            ('scheduledtasks', 'Display scheduled tasks (detailed)'),
-            ('deletedfiles', 'Display deleted files (Recycle Bin + Registry)'),
-            ('recentdocs', 'Display recently opened documents'),
-            ('userassist', 'Display UserAssist program execution tracking'),
-            ('persistence', 'Display persistence mechanisms (startup/registry/tasks/EXEs)'),
-            ('info', 'Display system information'),
-            ('privileges', 'List process privileges'),
-            ('getsids', 'Display process SIDs'),
-            ('threads', 'List process threads'),
-            ('filescan', 'Scan for file objects'),
-            ('envars', 'Display environment variables'),
-            ('search', 'Search for processes'),
-            # NEW PLUGINS
-            ('sandbox', 'Check if system is running in VM/sandbox'),
-            ('prefetch', 'Display Prefetch files information'),
-            ('shellbags', 'Display Shellbags information'),
-            ('amcache', 'Display Amcache.hve information'),
-            ('shimcache', 'Display Shimcache information'),
-            ('jumplists', 'Display Jump List information'),
-            ('timeliner', 'Create timeline forensics analysis'),
-            ('disklist', 'Display hard drives and storage devices'),
-            ('fileless', 'Advanced fileless malware and memory threat detection'),
-            ('clear', 'Clear screen'),
-            ('help', 'Show help'),
-            ('exit', 'Exit shell'),
-        ]
-        
-        print("\n" + "=" * 80)
-        print("                  Available Plugins")
-        print("=" * 80 + "\n")
-        
-        max_name_len = max(len(name) for name, _ in plugins)
-        
-        for plugin_name, description in plugins:
-            print(f"  • {plugin_name:<{max_name_len}}  -  {description}")
-        
-        print("\n" + "=" * 80)
-        print("Type 'help <plugin>' for detailed information about a plugin")
-        print("=" * 80 + "\n")
-    
-    def do_clear(self, arg):
-        """Clear the screen"""
-        self.context.clear_screen()
-    
-    def do_help(self, arg):
-        """Show help information
-        Usage: help [command]"""
-        if arg:
-            super().do_help(arg)
-        else:
-            self.do_plugins(arg)
-    
-    def do_exit(self, arg):
-        """Exit the live shell"""
-        return True
-    
-    def do_quit(self, arg):
-        """Exit the live shell"""
-        return True
-    
-    def do_EOF(self, arg):
-        """Exit on Ctrl+D"""
-        return True
-    
-    def emptyline(self):
-        """Do nothing on empty line"""
-        pass
-    
-    def default(self, line):
-        """Handle unknown commands"""
-        print(f"\n[ERROR] Unknown command: '{line}'")
-        print("Type 'help' or 'plugins' for available commands\n")
-    
-    def _parse_args(self, arg):
-        """Parse command line arguments"""
-        args = {}
-        parts = arg.split()
-        
-        i = 0
-        while i < len(parts):
-            if parts[i].startswith('--'):
-                key = parts[i]
-                if i + 1 < len(parts) and not parts[i + 1].startswith('--'):
-                    args[key] = parts[i + 1]
-                    i += 2
-                else:
-                    args[key] = True
-                    i += 1
-            else:
-                i += 1
-        
-        return args
-
-
-class LiveShell(interfaces.plugins.PluginInterface):
-    """Interactive shell for live system analysis without memory dumps."""
-    
-    _required_framework_version = (2, 0, 0)
-    _version = (2, 0, 0)
-    
-    @classmethod
-    def get_requirements(cls) -> List[interfaces.configuration.RequirementInterface]:
-        return [
-            requirements.BooleanRequirement(
-                name="interactive",
-                description="Launch interactive shell (default: True)",
-                optional=True,
-                default=True
-            ),
-        ]
-    
-    def _generator(self):
-        """Generator for non-interactive mode."""
-        
-        if not LIVE_ANALYSIS_AVAILABLE:
-            yield (0, ("Error: Required dependencies not available",))
-            yield (0, ("Install: pip install psutil pywin32",))
-            return
-        
-        return
-        yield
-    
-    def run(self):
-        """Run the live shell."""
-        
-        if not LIVE_ANALYSIS_AVAILABLE:
-            print("\n" + "=" * 80)
-            print("[ERROR] Live analysis dependencies not available!")
-            print("=" * 80)
-            print("\nPlease install required packages:")
-            print("  pip install psutil pywin32")
-            print("\nNote: Run as Administrator for full system access")
-            print("=" * 80 + "\n")
-            return renderers.TreeGrid([("Status", str)], self._generator())
-        
-        interactive = self.config.get("interactive", True)
-        
-        if interactive:
-            try:
-                shell = LiveShellCommand()
-                shell.cmdloop()
-            except KeyboardInterrupt:
-                pass
-            except Exception as e:
-                vollog.error(f"Error in live shell: {e}")
-        
-        return renderers.TreeGrid([("Status", str)], iter([]))
